@@ -23,14 +23,16 @@ const container = css`
 
 class Balance extends Component {
     render() {
+        const { isMobile, name, network, balance } = this.props
         return (
-            <div className={container + " qt-font-light" + (this.props.isMobile ? " mobile" : "")}>
-                Wallet Balance <Link to="exchange/wallets" className="ml-2"><img src={devicePath("public/images/open-icon.svg")} /></Link>
+            <div className={container + " qt-font-light" + (isMobile ? " mobile" : "")}>
+                Wallet Balance {name ? <Link to={"/" + network + "/wallets"} className="ml-2"><img src={devicePath("public/images/open-icon.svg")} /></Link> : null }
                 <div className="no-scroll-bar">
                     <div>
                         <table className="w-100 mt-3">
                             <tbody>
-                                {this.props.balance.map(row => {
+                                {Object.keys(balance).map(coin => {
+                                    const row = balance[coin]
                                     return (
                                         <tr key={row.asset}>
                                             <td><SymbolToken name={window.assets[row.asset].symbol} /></td>
@@ -49,7 +51,9 @@ class Balance extends Component {
 
 const mapStateToProps = (state) => ({
     isMobile: state.app.isMobile,
-    balance: state.app.balance || []
+    balance: state.app.balance || {},
+    network: state.app.network,
+    name: state.app.name
 });
 
 export default connect(mapStateToProps)(Balance);
