@@ -1,11 +1,15 @@
 import React, { Component } from 'react';
 import { css } from 'emotion'
 import { connect } from 'react-redux'
-import { LOGIN, switchTicker } from '../redux/actions/app.jsx'
-import { PrivateKey, changeWalletPassword, decryptWallet, encryptWallet } from "@quantadex/bitsharesjs";
+import { AccountLogin, ConnectAccount, GetAccount, TOGGLE_CONNECT_DIALOG, LOGOUT } from '../redux/actions/app.jsx'
+import { PrivateKey, PublicKey, decryptWallet, encryptWallet, hash } from "@quantadex/bitsharesjs";
 import WalletApi from "../common/api/WalletApi";
-import QTTabBar from './ui/tabBar.jsx'
 import Loader from '../components/ui/loader.jsx'
+import Lock from './ui/account_lock.jsx'
+import CONFIG from '../config.js'
+import bs58 from 'bs58'
+import Recaptcha from 'react-google-invisible-recaptcha';
+import {getItem, setItem, removeItem, clear } from "../common/storage.js";
 
 const container = css`
     text-align: center;
@@ -18,13 +22,12 @@ const container = css`
         background-color: transparent;
         border: 2px solid #66d7d7;
         padding: 10px 30px;
-        border-radius: 20px;
+        border-radius: 30px;
         margin: 10px;
         cursor: pointer;
     }
 
     .connect-link {
-        display: inline;
         color: #fff;
         text-decoration: underline;
         cursor: pointer;
@@ -39,22 +42,24 @@ const container = css`
     }
     &.link.mobile {
         background: none;
+        font-size: 12px;
+        padding: 0;
+        margin: 0;
     }
 
     &.mobile  {
-        padding: 0;
-        margin: 0;
+        font-size: 16px;
     }
 
 `
 
 const dialog = css`
-    position: absolute;
+    position: fixed;
     top: 0;
     bottom: 0;
     left: 0;
     right: 0;
-    background-color: rgba(0,0,0,0.6);
+    background-color: rgba(0,0,0,0.8);
     font-size: 13px;
     z-index: 999;
 
@@ -63,7 +68,8 @@ const dialog = css`
     }
 
     .link {
-        color: #66d7d7;
+        text-align: right;
+        color: #66d7d7 !important;
         cursor: pointer;
     }
 
@@ -74,7 +80,7 @@ const dialog = css`
 
     .container {
         position: relative;
-        width: auto;
+        width: 100%;
         max-width: 750px;
         background-color: #4f637e;
         border-radius: 5px;
@@ -95,7 +101,6 @@ const dialog = css`
             position: relative;
             background-color: #fff;
             padding: 20px;
-            margin: 20px 0;
             color: #999;
             font-size: 15px;
 
@@ -108,8 +113,22 @@ const dialog = css`
                 border-radius: 4px;
             }
 
+            input:read-only {
+                background: #eee;
+            }
+
+            .generate-key {
+                background: url(${devicePath('public/images/generate.svg')}) no-repeat 0 50%;
+                padding-left: 18px;
+            }
+
+            .personal-key input {
+                padding: 0;
+                height: auto;
+                width: auto ;
+            }
+
             .error {
-                position: absolute;
                 color: #f0185c;
                 font-size: 11px;
             }
@@ -127,7 +146,54 @@ const dialog = css`
             button:disabled {
                 background-color: #999;
             }
+
+            select {
+                border: 1px solid #ccc;
+                border-radius: 3px;
+            }
         }
+    }
+
+    .selection-tab {
+        align-items: center;
+        padding: 7px 20px;
+        background: #d8d8d8;
+        color: #000;
+        cursor: pointer;
+        font-size: 14px;
+
+        .select-btn {
+            border: 3px solid #2ebeed;
+            border-radius: 100%;
+            margin-right: 12px;
+
+            div {
+                width: 16px;
+                height: 16px;
+                border: 3px solid #fff;
+                border-radius: 100%;
+                background: #fff;
+            }
+        }
+    }
+
+    .active {
+        .selection-tab {
+            background: #fff;
+            cursor: default;
+
+            .select-btn div {
+                background: #2ebeed;
+            }
+        }
+    }
+
+    .referral {
+        margin-top: -15px;
+    }
+
+    .container.testnet {
+        background-color: #555;
     }
 
     .warning {
@@ -188,9 +254,13 @@ const dialog = css`
     }
 
     &.mobile {
+        position: relative;
+        background-color: transparent;
+        z-index: 1;
+
         .container {
             width: 100% !important;
-
+            background-color: transparent;
             .input-container {
                 button {
                     margin: 0;
@@ -201,31 +271,44 @@ const dialog = css`
     }
 `
 
-class ConnectLink extends Component {
-    render() {
-        return (
-            <div className={container + " link cursor-pointer" + (this.props.isMobile ? " mobile" : "")} onClick={() => this.props.onOpen("connect")}>CONNECT WALLET</div>
-        )
-    }
-    
-} 
-
 class Connect extends Component {
+    openDialog(dialogType) {
+        const { mobile_nav, dispatch } = this.props
+        if (mobile_nav) {
+            mobile_nav(dialogType)
+        } else {
+            dispatch({
+                type: TOGGLE_CONNECT_DIALOG,
+                data: dialogType
+            })
+        }
+    }
+
     render() {
+        const { type, isMobile } = this.props
         return (
-            <div className={container}>
-                <p>Connect your <span className="qt-font-bold">Quanta</span> wallet to start trading in this market.</p>
-                <button onClick={() => this.props.onOpen("create")}>GET STARTED</button>
-                <div>
-                    or<br/>
-                    <div className="connect-link" onClick={() => this.props.onOpen("connect")}>Connect Wallet</div>
+            <React.Fragment>
+                {
+                    type == "link" ? 
+                    <div className={container + " link cursor-pointer" + (isMobile ? " mobile" : "")} onClick={() => this.openDialog("connect")} >CONNECT WALLET</div>
+                :   type == "lock" ?
+                    <Lock unlock={() => this.openDialog("connect")}/>
+                    :
+                    <div className={container + (isMobile ? " mobile" : "")}>
+                        <p>Connect your <span className="qt-font-bold">Quanta</span> wallet to start trading in this market.</p>
+                        <button onClick={() => this.openDialog("create")}>CREATE WALLET</button>
+                        <div>
+                            or
+                            <div className="connect-link mt-3" onClick={() => this.openDialog("connect")}>Connect Wallet</div>
+                        </div>
                 </div>
-            </div>
+                }
+            </React.Fragment>
         )
     }
 }
 
-class ConnectDialog extends Component {
+export class ConnectDialog extends Component {
     constructor(props) {
         super(props);
         this.state = {
@@ -233,8 +316,13 @@ class ConnectDialog extends Component {
             selectedTabIndex: 0,
             regStep: 1,
             encryptStep: 0,
-            encrypted_data: {},
+            no_email: false,
+            encrypted_data: null,
             private_key: "",
+            public_key: "",
+            valid_key: true,
+            email: "",
+            email_code: "",
             username: "",
             password: "",
             confirm_password: "",
@@ -242,29 +330,63 @@ class ConnectDialog extends Component {
             errorMsg: "",
             downloaded: false,
             uploaded_file_name: false,
+            scan_qr: true,
+            bip58: ""
         };
 
         this.handleChange = this.handleChange.bind(this)
-        this.KeyConnect = this.KeyConnect.bind(this)
         this.KeyCreate = this.KeyCreate.bind(this)
+        this.VerifyEmail = this.VerifyEmail.bind(this)
+        this.ConfirmEmail = this.ConfirmEmail.bind(this)
+        this.Register = this.Register.bind(this)
         this.KeyDownload = this.KeyDownload.bind(this)
         this.ConnectEncrypted = this.ConnectEncrypted.bind(this)
         this.ConnectPrivateKey = this.ConnectPrivateKey.bind(this)
+        this.AccountSelect = this.AccountSelect.bind(this)
         this.EncryptKey = this.EncryptKey.bind(this)
-        this.DownloadKey = this.DownloadKey.bind(this)
+        this.downloadKey = this.downloadKey.bind(this)
+        this.generateKey = this.generateKey.bind(this)
     }
 
+    componentDidMount() {
+        this.generateKey()
+        this.loadStore().then(() => {            
+        })
+
+        const search = window.location.search.slice(1).split("=")
+        const referrer = search.indexOf("referrer") !== -1 && search[search.indexOf("referrer") + 1]
+        if (!referrer) return
+        
+        this.setState({referrer})
+        
+        GetAccount(referrer).then(e => {
+            if (e.membership_expiration_date !== "1969-12-31T23:59:59") {
+                this.setState({referrer_error: "Referrer account is not a lifetime member - user need to activate the referral program"})
+            }
+        }).catch(error => {
+            this.setState({referrer_error: "Referrer account does not exist"})
+        })
+    }
 
     componentWillReceiveProps(nextProps) {
-        if (this.state.dialogType !== nextProps.default) {
+        if (this.props.default !== nextProps.default && !this.props.isMobile) {
             this.setState({dialogType: nextProps.default})
         }
+    }
+
+    static getDerivedStateFromError(error) {
+        // Update state so the next render will show the fallback UI.
+        console.log(error)
+    }
+    componentDidCatch(error, info) {
+        console.log(error, info)
     }
 
     resetInputs(e = {}) {
         this.setState({
             ...e,
             private_key: "",
+            bip58: "",
             username: "",
             password: "",
             confirm_password: "",
@@ -274,8 +396,10 @@ class ConnectDialog extends Component {
     }
 
     closeDialog() {
-        document.getElementById("connect-dialog").style.display = "none"
-        this.resetInputs()
+        this.props.dispatch({
+            type: TOGGLE_CONNECT_DIALOG,
+            data: false
+        })
     }
     
     handleSwitch(index) {
@@ -303,52 +427,92 @@ class ConnectDialog extends Component {
         return pw1.length >= 8 && pw1.match(/[A-Z]/) && pw1.match(/[0-9]/)
     }
 
-    ConnectWithBin() {
+    ConnectWithBin(type = undefined, nav = true) {
+        const { mobile_nav, dispatch } = this.props
+        const { password, bip58, storeEncrypted } = this.state
         try {
-			const decrypted = decryptWallet(this.state.encrypted_data, this.state.password)
+            var encrypted_data
+            if (type === "bip58") {
+                const hex = bs58.decode(bip58).toString('hex')
+                if (hex.length !== 192) throw "Invalid Key"
+                encrypted_data = {encryption_key: hex.slice(0,96), wallet_encryption_key: hex.slice(-96)}
+            } else if (type === "qr") {
+                encrypted_data = JSON.parse(atob(bip58))
+            } else {
+                encrypted_data = this.state.encrypted_data || JSON.parse(storeEncrypted)
+            }
+            
+			const decrypted = decryptWallet(encrypted_data, password)
             const private_key = decrypted.toWif()
-			this.props.dispatch({
-                type: LOGIN,
-                private_key: private_key
-            });
-            this.props.dispatch(switchTicker(this.props.currentTicker))
+
+            dispatch(AccountLogin(private_key)).then((e) => {
+                setItem("encrypted_data", JSON.stringify(encrypted_data))
+
+                if(Array.isArray(e)) {
+                    this.setState({account_select: true, accounts_list: e, private_key})
+                } else if (nav && mobile_nav) {
+                    mobile_nav()
+                }
+            })
+            .catch(error => {
+                console.log(error)
+                this.setState({authError: true, errorMsg: error})
+            })
+
 		} catch(e) {
-			console.log(e)
-			this.setState({authError: true, errorMsg: "Incorrect Password"})
+            console.log(e)
+            let errorMsg
+            if (e == "Invalid Key") {
+                errorMsg = e
+            } else {
+                errorMsg = "Your password and key does not match"
+            }
+			this.setState({authError: true, errorMsg })
 		}
         
     }
     
     ConnectWithKey() {
+        const { mobile_nav, dispatch } = this.props
+        const { private_key } = this.state
         try {
-			const pKey = PrivateKey.fromWif(this.state.private_key);
-			this.props.dispatch({
-                type: LOGIN,
-                private_key: this.state.private_key
-            });
-            this.props.dispatch(switchTicker(this.props.currentTicker))
+            const pKey = PrivateKey.fromWif(private_key);
+
+            dispatch(AccountLogin(private_key)).then((e) => {
+                removeItem("encrypted_data")
+
+                if(Array.isArray(e)) {
+                    this.setState({account_select: true, accounts_list: e})
+                } else if (mobile_nav) {
+                    mobile_nav()
+                }
+            })
+            .catch(error => {
+                this.setState({authError: true, errorMsg: error})
+            })
 		} catch(e) {
-			console.log(e)
 			this.setState({authError: true, errorMsg: "Invalid Key"})
 		}
     }
 
     Encrypt() {
-        if (this.state.password !== this.state.confirm_password) {
+        const { network } = this.props
+        const { password, confirm_password, private_key, username } = this.state
+        if (password !== confirm_password) {
             this.setState({authError: true, errorMsg: "Your password inputs are not the same"})
             return
         } 
 
-        if (!this.validatePassword(this.state.password)) {
+        if (!this.validatePassword(password)) {
             this.setState({authError: true, errorMsg: "Password must contains at least 8 characters, 1 uppercase, and 1 number."})
             return
         }
 
         try {
-            const key = PrivateKey.fromWif(this.state.private_key)
-            const encryption = encryptWallet(key, this.state.password)
+            const key = PrivateKey.fromWif(private_key)
+            const encryption = encryptWallet(key, password)
             const text= JSON.stringify(encryption)
-            this.download("quanta_wallet.json", text)
+            this.download(`quanta_${network}_${username}.json`, text)
             this.setState({downloaded: true, authError: false})
         } catch(e) {
             this.setState({authError: true, errorMsg: "Invalid Key"})
@@ -356,70 +520,90 @@ class ConnectDialog extends Component {
     }
 
     registerAccount() {
-        if (this.state.password !== this.state.confirm_password) {
-            this.setState({authError: true, errorMsg: "Your password inputs are not the same"})
-            return
-        } 
+        const { password, confirm_password, username, personal_key, public_key, generated_private_key, no_email, email, email_code, referrer, referrer_error } = this.state
+        if(!personal_key) {
+            if (password !== confirm_password) {
+                //this.recaptcha.reset()
+                this.setState({authError: true, errorMsg: "Your password inputs are not the same"})
+                return
+            } 
 
-        if (!this.validatePassword(this.state.password)) {
-            this.setState({authError: true, errorMsg: "Password must contains at least 8 characters, 1 uppercase, and 1 number."})
-            return
-        }
-
-        const keys = WalletApi.generate_key()
-
-		this.setState({processing: true, private_key: keys.privateKey})
-
-		fetch("/api/register", {
-			method: "post",
-			headers: {
-					"Content-Type": "application/json",
-					Accept: "application/json"
-			},
-			body: JSON.stringify({
-				user_name: this.state.username.toLowerCase(),
-				public_key: keys.publicKey,
-			})
-		}).then(response => {
-			if (response.status == 200) {
-				this.setState({
-                    regStep:2,
-					authError: false
-				});
-				return response.json();
-			} else {
-				return response.json().then(res => {
-					var msg;
-					if (res.message.includes("already exists")) {
-						msg = "Username already exist"
-					} else if (res.message.includes("is_valid_name")) {
-						msg = "Name must start with a letter and only contains alpha numeric, dash, and dot"
-					} else {
-						msg = "Server error. Please try again."
-					}
-					this.setState({
-						authError: true,
-						errorMsg: msg
-					});
-				});
-			}
-		})
-		.finally(() => {
-			this.setState({processing: false})
-		})
-    }
-
-    DownloadKey() {
-        const checks = document.forms.agreements.getElementsByTagName("input")
-        for (let input of checks) {
-            if (!input.checked) {
-                this.setState({authError: true, errorMsg: "You must agree to all of the above"})
+            if (!this.validatePassword(password)) {
+                //this.recaptcha.reset()
+                this.setState({authError: true, errorMsg: "Password must contains at least 8 characters, 1 uppercase, and 1 number."})
                 return
             }
         }
+        
+        this.setState({processing: true, authError: false})
 
+        const encryption = encryptWallet(PrivateKey.fromWif(generated_private_key), password)
+        const encrypted_data= JSON.stringify(encryption)
+        var reg_json = {}
+
+        if (no_email) {
+            reg_json = {
+                name: username.toLowerCase(),
+                public_key: public_key,
+            }
+        } else {
+            reg_json = {
+                email: email,
+                confirm: email_code,
+                public_key: public_key,
+                account: username.toLowerCase(),
+                json: window.btoa(encrypted_data),
+            }
+        }
+
+        if (referrer && !referrer_error) {
+            reg_json.referrer= referrer
+        }
+
+        fetch(CONFIG.getEnv().API_PATH + (no_email ? "/register_account" : "/send_walletinfo"), {
+            method: "post",
+            mode: "cors",
+            headers: {
+                "Content-Type": "application/json",
+                Accept: "application/json"
+            },
+            body: JSON.stringify(reg_json)
+        }).then(e => e.json()).then(response => {
+            if (response.success || response.status == "success") {
+                this.setState({
+                    regStep: 4,
+                    authError: false,
+                    private_key: generated_private_key
+                });
+
+                if (window.isApp) {
+                    this.setState({encrypted_data: JSON.parse(encrypted_data)})
+                    this.ConnectWithBin(undefined, false)
+                }
+            } else {
+                //this.recaptcha.reset()
+                let error = response.error || response.message
+                var msg;
+                if (error.includes("already exists")) {
+                    msg = "Username already exist"
+                } else if (error.includes("is_valid_name")) {
+                    msg = "Name must start with a letter and only contains alpha numeric, dash, and dot"
+                } else {
+                    msg = error || "Server error. Please try again."
+                }
+                this.setState({
+                    authError: true,
+                    errorMsg: msg
+                });
+            }
+        }).finally(() => {
+            this.setState({processing: false})
+        })
+    }
+
+    downloadKey() {
         this.Encrypt()
-        this.resetInputs({downloaded: true})
+        this.setState({downloaded: true})
     }
 
     uploadFile(file) {
@@ -442,41 +626,439 @@ class ConnectDialog extends Component {
         this.uploadFile(file)
     }
 
+    VerifyEmail() {
+        const self = this
+        const { email, email_error, processing } = this.state
+
+        function validateEmail() {
+            const isValid = email.match(/^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/)
+            
+            if (!window.isApp) {
+                if (!isValid) {
+                    this.recaptcha.reset()
+                    self.setState({ email_error: "Not a valid email address." })
+                } else {
+                    this.recaptcha.execute()
+                }
+            } else {
+                if (!isValid) {
+                    self.setState({ email_error: "Not a valid email address." })
+                } else {
+                    verify(true)
+                }
+            }
+        }
+
+        function verify(sign) {
+            const { email, email_error, processing } = self.state
+            var body = null;
+            self.setState({email_error: false, processing: true})
+            var sig = {}
+            if (sign === true) {
+                body = JSON.stringify({ email: email })
+                sig["signature"] = hash.HmacSHA256(body, k).toString('hex');
+            } else {
+                body = JSON.stringify({ email: email, recaptcha: this.recaptcha.getResponse() })
+            }
+
+            fetch(CONFIG.getEnv().API_PATH + "/verify_email", {
+                mode: "cors",
+                method: "post",
+                headers: {
+                        "Content-Type": "application/json",
+                        Accept: "application/json",
+                        ...sig
+                },
+                body: body
+            }).then(e => e.json()).then(response => {
+                if (response.success) {
+                    self.setState({
+                        regStep:2
+                    });
+                } else {
+                    self.setState({
+                        email_error: response.message || response.error || "Server error. Please try again."
+                    });
+                }
+            }).finally(() => {
+                self.setState({processing: false})
+            })
+        }
+        
+        return (
+            <React.Fragment>
+                <h5 className="text-dark">1/3 - VERIFY YOUR EMAIL</h5>
+                <p className="qt-font-small my-4">
+                    Encrypted wallet via email protects from accidental loss, and provide quick access to our mobile app.  
+                    The private key is generated locally and encrypted with your password.
+                </p>
+
+                <label>EMAIL TO RECEIVE THE ENCRYPTED KEY</label><br/>
+                <input type="email" name="email" placeholder="Email" autoFocus spellCheck="false"
+                    value={email} onChange={(e) => this.setState({email: e.target.value})}
+                />
+                {email_error ? <span className="text-danger small">{email_error}</span> : null}
+                <div className="text-center mt-5">
+                    <button onClick={validateEmail.bind(this)} disabled={processing || !email} >
+                        {processing ? <Loader /> : "VERIFY EMAIL"}
+                    </button>
+                    {!window.isApp && <Recaptcha
+                        ref={ ref => this.recaptcha = ref }
+                        sitekey="6Lc4OZ4UAAAAAEfECNb09tkSL_3UBCuV_sdITK5B"
+                        onResolved={ verify.bind(this) } />
+                    }
+                </div>
+                {window.isApp ? null :
+                    <div className="link qt-font-small text-center mt-4" 
+                        onClick={() => this.setState({no_email: true, regStep: 3})}>
+                        <u>Skip. Download locally</u>
+                    </div>
+                }
+            </React.Fragment>
+        )
+    }
+
+    ConfirmEmail() {
+        const self = this
+        const { email, email_code, email_code_error, processing } = this.state
+        function confirm() {
+            self.setState({processing: true, email_code_error: false})
+            fetch(CONFIG.getEnv().API_PATH + "/confirm_email", {
+                mode: "cors",
+                method: "post",
+                mode: "cors",
+                headers: {
+                        "Content-Type": "application/json",
+                        Accept: "application/json"
+                },
+                body: JSON.stringify({ email: email, confirm: email_code })
+            }).then(e => e.json()).then(response => {
+                if (response.success) {
+                    self.setState({
+                        regStep:3
+                    });
+                } else {
+                    self.setState({
+                        email_code_error: response.message || response.error || "Server error. Please try again."
+                    });
+                }
+            }).finally(() => {
+                self.setState({processing: false})
+            })
+        }
+
+        return (
+            <React.Fragment>
+                <h5 className="text-dark">2/3 - CONFIRM YOUR EMAIL</h5>
+                <p className="qt-font-small my-4">
+                    We sent a code to {email}. Enter it below:
+                </p>
+
+                <label>VERIFICATION CODE</label><br/>
+                <input type="text" name="email_code" spellCheck="false" autoFocus placeholder="Code from email..."
+                    value={email_code} onChange={(e) => this.setState({email_code: e.target.value})}
+                    onKeyPress={e => {
+                        if (e.key == "Enter" && !processing) {
+                            confirm()
+                        }
+                    }}
+                />
+                {email_code_error ? <span className="text-danger small">{email_code_error}</span> : null}
+
+                <div className="text-center mt-5">
+                    <button onClick={confirm} disabled={processing} >
+                        {processing ? <Loader /> : "NEXT"}
+                    </button>
+                </div>
+            </React.Fragment>
+        )
+    }
+
+    generateKey() {
+        const keys = WalletApi.generate_key()
+        this.setState({ public_key: keys.publicKey, generated_private_key: keys.privateKey, valid_key: true })
+    }
+
+    Register() {
+        const { no_email, username, public_key, personal_key, valid_key, password, confirm_password, authError, errorMsg, processing } = this.state
+
+        return (
+            <React.Fragment>
+                <h5 className="text-dark">{no_email ? "": "3/3 - "}SETUP YOUR WALLET</h5>
+                <div className="mb-2">
+                    <label>USERNAME</label>
+                    <br/>
+                    <input id="name-input" type="text" autoComplete="off" autoFocus placeholder="Username" spellCheck="false" 
+                        value={username} onChange={(e) => this.setState({username: e.target.value})}/>
+                </div>
+
+                <div className="mb-2">
+                    <div className="d-flex justify-content-between">
+                        <label>PUBLIC KEY</label>
+                        <div className="d-flex align-items-center">
+                            <label className="generate-key cursor-pointer mb-0" onClick={this.generateKey}>Generate</label>
+                            { no_email ?
+                                <div className="personal-key d-flex align-items-center ml-5">
+                                    <input type="checkbox" id="personal-key" name="personal-key" 
+                                        onChange={e => {
+                                            this.setState({personal_key: e.target.checked})
+                                            if (!e.target.checked) this.generateKey()
+                                        }}/>
+                                    <label className="mb-0 ml-2" htmlFor="personal-key">Use my own public key</label>
+                                </div>
+                                : null
+                            }
+                        </div>
+                    </div>
+                    <input type="text" autoComplete="off" placeholder="Public Key" spellCheck="false" readOnly={!personal_key}
+                        value={public_key} onChange={(e) => {
+                            const valid_key = PublicKey.fromPublicKeyString(e.target.value)
+                            this.setState({public_key: e.target.value, valid_key: valid_key && true})
+                        }}/>
+                    {!valid_key ? <span className="text-danger small">Invalid Key</span> : null}
+                </div>
+                {!personal_key ?
+                    <React.Fragment>
+                        <div className="mb-2">
+                            <label>WALLET PASSWORD</label>
+                            <br/>
+                            <input id="pw-input" type="password" placeholder="Password"
+                                value={password} onChange={(e) => this.setState({password: e.target.value})}/>
+                        </div>
+
+                        <div className="mb-2">
+                            <label>CONFIRM WALLET PASSWORD</label>
+                            <br/>
+                            <input id="pwconf-input" type="password" placeholder="Confirm Password" spellCheck="false" 
+                                value={confirm_password} onChange={(e) => this.setState({confirm_password: e.target.value})}/>
+                        </div>
+                    </React.Fragment>
+                : null
+                }
+
+                <span className="error" hidden={!authError}>{errorMsg}</span><br/>
+
+                <div className="text-center">
+                    <button onClick={this.registerAccount.bind(this)} 
+                        disabled={!valid_key || username.length == 0 || (!personal_key && (password.length == 0 || confirm_password.length == 0)) || processing}>
+                        {processing ? <Loader /> : "REGISTER ACCOUNT"}
+                    </button>
+                </div>
+            </React.Fragment>
+        )
+    }
+
+    keyCreateStep(step) {
+        switch (step) {
+            case 1:
+                return <this.VerifyEmail />
+            case 2:
+                return <this.ConfirmEmail />
+            case 3:
+                return <this.Register />
+            case 4:
+                return <this.KeyDownload />
+        }
+    }
+
+    async loadStore() {
+        try {
+            const storeName = await getItem("name")
+            const storeId = await getItem("id")
+            const storeEncrypted = await getItem("encrypted_data")
+            this.setState({storeName, storeEncrypted, storeId})
+        } catch(e) {
+            console.log(e)
+        }
+    }
+
     ConnectEncrypted() {
+        const self = this
+        const { isMobile, network, dispatch } = this.props
+        const { encrypted_data, uploaded_file_msg, password, authError, errorMsg, scan_qr, bip58, storeName, storeEncrypted, processing } = this.state
+
+        if (window.isApp) {
+            function scanQR() {
+                self.setState({processing: true})
+                cordova.plugins.barcodeScanner.scan(
+                    function (result) {
+                        if (result.text) {
+                            self.setState({bip58: result.text, processing: false})
+                        }
+                    },
+                    function (error) {
+                        alert("Scanning failed: " + error);
+                        self.setState({processing: false})
+                    },
+                    {
+                        preferFrontCamera: false, // iOS and Android
+                        showFlipCameraButton: true, // iOS and Android
+                        showTorchButton: true, // iOS and Android
+                        torchOn: false, // Android, launch with the torch switched on (if available)
+                        saveHistory: false, // Android, save scan history (default false)
+                        prompt: "Place QUANTA QRCode inside the scan area", // Android
+                        resultDisplayDuration: 500, // Android, display scanned text for X ms. 0 suppresses it entirely, default 1500
+                        formats: "QR_CODE", // default: all but PDF_417 and RSS_EXPANDED
+                        orientation: "portrait", // Android only (portrait|landscape), default unset so it rotates with the device
+                        disableAnimations: true, // iOS
+                        disableSuccessBeep: true // iOS and Android
+                    }
+                );                                
+            }
+            return (
+                storeEncrypted ?
+                <div className="input-container text-center">
+                    <div className="text-secondary text-center mb-2">
+                        Continue as <b>{storeName}</b> or&nbsp;
+                        <span className="qt-color-theme"
+                            onClick={() => {
+                                const c = confirm("This will remove your credentials from current device. Make sure you have backup of you Private Key before continue!")
+                                if (c) {
+                                    this.setState({storeEncrypted: null, storeName: null})
+                                    dispatch({
+                                        type: LOGOUT
+                                    })
+                                }
+                            }}
+                        >
+                            Logout
+                        </span>
+                    </div>
+                    <div className="text-left">
+                        <label>PASSWORD</label><br/>
+                        <input type="password" name="password" placeholder="Password" 
+                            value={password} onChange={(e) => this.setState({password: e.target.value})}
+                            onKeyPress={e => {
+                                if (e.key == "Enter" && password.length >= 8) {
+                                    this.ConnectWithBin()
+                                }
+                            }}
+                            />
+                    </div>
+                    <span className="text-danger small">{errorMsg}</span>
+                    <button className="mt-5" 
+                        disabled={password.length < 8}
+                        onClick={() => this.ConnectWithBin()}>Connect Wallet</button>
+                </div>
+                    
+                : 
+                scan_qr ? 
+                    bip58 ?
+                    <div className="input-container">
+                        <div className="d-inline-block" onClick={() => this.setState({bip58: ""})}>
+                            <img src={devicePath("public/images/back-btn-black.svg")} />
+                        </div>
+                        <p className="mb-5 mt-4">
+                            Key from QR Code: {bip58.slice(0,6)}.....{bip58.slice(-6)}
+                        </p>
+                        <div className="text-left">
+                            <label>PASSWORD</label><br/>
+                            <input type="password" name="password" placeholder="Password" 
+                                value={password} onChange={(e) => this.setState({password: e.target.value})}
+                                onKeyPress={e => {
+                                    if (e.key == "Enter" && password.length > 0) {
+                                        this.ConnectWithBin("qr")
+                                    }
+                                }}
+                            />
+                        </div>
+                        <span className="text-danger small">{errorMsg}</span>
+                        <button className="mt-5" 
+                            disabled={password.length < 8}
+                            onClick={() => this.ConnectWithBin("qr")}>Connect Wallet</button>
+                    </div>
+                    :
+                    <div className="input-container text-center">
+                        <p className="qt-font-small mb-5">
+                            On your desktop, open the email containing your wallet QR code.
+                        </p>
+
+                        <button disabled={processing} onClick={scanQR.bind(this)}>{processing ? <Loader /> : "SCAN QR CODE"}</button>
+                        <div className="mt-4 mb-5" onClick={() => this.resetInputs({scan_qr: !scan_qr})}>
+                            Or <span className="qt-color-theme">Copy & Paste Base58 Key</span>
+                        </div>
+                    </div>
+                :
+                    <div className="input-container text-center">
+                        <p className="qt-font-small mb-5">
+                            On your desktop, open the email containing your wallet BIP58 encrypted key.
+                        </p>
+                        <div className="text-left">
+                            <label>BIP58 ENCRYPTED KEY</label><br/>
+                            <input type="text" name="bip58" placeholder="BIP58 Encrypted Key" 
+                                value={bip58} onChange={(e) => this.setState({bip58: e.target.value})}
+                            />
+                        </div>
+                        <div className="text-left">
+                            <label>PASSWORD</label><br/>
+                            <input type="password" name="password" placeholder="Password" 
+                                value={password} onChange={(e) => this.setState({password: e.target.value})}
+                                onKeyPress={e => {
+                                    if (e.key == "Enter" && password.length >= 8 && bip58) {
+                                        this.ConnectWithBin("bip58")
+                                    }
+                                }}
+                                />
+                        </div>
+                        <span className="text-danger small">{errorMsg}</span>
+                        <button className="mt-5" 
+                            disabled={password.length < 8 || !bip58}
+                            onClick={() => this.ConnectWithBin("bip58")}>Connect Wallet</button>
+                        <div className="mt-4" onClick={() => this.resetInputs({scan_qr: !scan_qr})}>
+                            Or <span className="qt-color-theme">Scan QR Code</span>
+                        </div>
+                    </div>
+            )
+        }
+
         return (
             <div className="input-container">
-                <div className={"drop-zone align-items-center" + (this.props.isMobile ? " pt-3" : " d-flex")} onDragOver={(e)=> e.preventDefault()} onDrop={(e) => this.handleDrop(e)}>
-                    Drop your backup file in this area or&nbsp;<label htmlFor="file">browse your files.</label>
+                {!encrypted_data && storeEncrypted ?
+                <div className="text-secondary text-center mb-2">
+                    Continue as <span className="qt-color-theme">{storeName}</span> or&nbsp;<label className="cursor-pointer" htmlFor="file"><u>browse your files.</u></label>
                     <input className="d-none" type="file" name="file" id="file" accept=".json" onChange={(e) => this.uploadFile(e.target.files[0])}/>
                 </div>
+                : 
+                <React.Fragment>
+                    <div className={"drop-zone align-items-center" + (isMobile ? " pt-3" : " d-flex")} onDragOver={(e)=> e.preventDefault()} onDrop={(e) => this.handleDrop(e)}>
+                        Drop your backup file in this area or&nbsp;<label htmlFor="file">browse your files.</label>
+                        <input className="d-none" type="file" name="file" id="file" accept=".json" onChange={(e) => this.uploadFile(e.target.files[0])}/>
+                    </div>
+                    
+                    <div className="d-flex justify-content-between qt-font-small mb-2">
+                        <div>{uploaded_file_msg}</div>
+                        <div className="link text-right"
+                            onClick={() => this.resetInputs({encryptStep: 1})}>Convert Private Key</div>
+                    </div>
+                </React.Fragment>
+                }
                 
-                <div className="d-flex justify-content-between qt-font-small mb-2">
-                    <div>{this.state.uploaded_file_msg}</div>
-                    <div className="link text-right"
-                        onClick={() => this.resetInputs({encryptStep: 1})}>I don’t have a .json-file</div>
-                </div>
 
                 <label>PASSWORD</label><br/>
                 <input type="password" name="password" placeholder="Password" 
-                    value={this.state.password} onChange={(e) => this.setState({password: e.target.value})}
+                    value={password} onChange={(e) => this.setState({password: e.target.value})}
                     onKeyPress={e => {
-                        if (e.key == "Enter" && this.state.password.length > 0) {
+                        if (e.key == "Enter" && password.length > 0) {
                             this.ConnectWithBin()
                         }
                        }}
                     /><br/>
-                <span className="error" hidden={!this.state.authError}>{this.state.errorMsg}</span><br/>
+                <span className="error" hidden={!authError}>{errorMsg}</span><br/>
                 <div className="text-center">
-                    <button onClick={this.ConnectWithBin.bind(this)} disabled={this.state.password.length < 8}>Connect Wallet</button>
+                    <button onClick={this.ConnectWithBin.bind(this)} 
+                        disabled={password.length < 8 || !(encrypted_data || storeEncrypted)}>
+                        Connect Wallet
+                    </button>
                 </div>
             </div>
         )
     }
 
     EncryptKey() {
+        const { private_key, password, confirm_password, authError, errorMsg, downloaded } = this.state
         return (
             <div className="input-container">
-                <div className="link float-right qt-font-small" onClick={() => this.resetInputs({encryptStep: 0})}>Close</div>
+                <div className="link float-right qt-font-small" onClick={() => this.resetInputs({encryptStep: 0})}>Back</div>
                 <h5>CREATE AN ENCRYPTED PRIVATE "JSON" KEY</h5>
                 <p className="info">
                     Encrypting your private key will make it safer to login, and store.
@@ -486,30 +1068,30 @@ class ConnectDialog extends Component {
                 <div className="mb-2">
                     <label>PRIVATE KEY</label><br/>
                     <input id="key-input" type="text" autoComplete="off" placeholder="Private Key" spellCheck="false" 
-                        value={this.state.private_key} onChange={(e) => this.setState({private_key: e.target.value})}/>
+                        value={private_key} onChange={(e) => this.setState({private_key: e.target.value})}/>
                 </div>
                 
                 <div className="mb-2">
                     <label>PASSWORD</label><br/>
                     <input id="en-pw-input" type="password" placeholder="Password"
-                        value={this.state.password} onChange={(e) => this.setState({password: e.target.value})}/>
+                        value={password} onChange={(e) => this.setState({password: e.target.value})}/>
                 </div>
 
                 <div className="mb-2">
                     <label>CONFIRM PASSWORD</label><br/>
                     <input id="en-pwconf-input" type="password" placeholder="Confirm Password" spellCheck="false" 
-                        value={this.state.confirm_password} onChange={(e) => this.setState({confirm_password: e.target.value})}/>
+                        value={confirm_password} onChange={(e) => this.setState({confirm_password: e.target.value})}/>
                 </div>
 
-                <span className="error" hidden={!this.state.authError}>{this.state.errorMsg}</span><br/>
+                <span className="error" hidden={!authError}>{errorMsg}</span><br/>
 
                 <div className="text-center">
-                    <button onClick={this.Encrypt.bind(this)} disabled={this.state.private_key.length == 0 || this.state.password.length == 0 || this.state.confirm_password.length == 0}>
+                    <button onClick={this.Encrypt.bind(this)} disabled={private_key.length == 0 || password.length == 0 || confirm_password.length == 0}>
                         ENCRYPT KEY
                     </button>
                 </div>
 
-                <div className={"link qt-font-small text-center" + (!this.state.downloaded ? " invisible" : "")} onClick={() => this.resetInputs({encryptStep: 0})}>
+                <div className={"link qt-font-small text-center" + (!downloaded ? " invisible" : "")} onClick={() => this.resetInputs({encryptStep: 0})}>
                     <u>Proceed to Connect Wallet</u>
                 </div>
             </div>
@@ -517,166 +1099,174 @@ class ConnectDialog extends Component {
     }
 
     ConnectPrivateKey() {
+        const { private_key, authError, errorMsg } = this.state
         return (
             <div className="input-container">
                 <label>PRIVATE KEY</label><br/>
                 <input id="pkey-input" type="text" autoComplete="off" autoFocus placeholder="Private Key" spellCheck="false" 
-                   value={this.state.private_key}
+                    name="privateKey"
+                   value={private_key}
                    onChange={(e) => this.setState({private_key: e.target.value})}
                    onKeyPress={e => {
-                    if (e.key == "Enter" && this.state.private_key.length > 0) {
+                    if (e.key == "Enter" && private_key.length > 0) {
                         this.ConnectWithKey()
                     }
                    }}/><br/>
-                <span className="error" hidden={!this.state.authError}>{this.state.errorMsg}</span><br/>
+                <span className="error" hidden={!authError}>{errorMsg}</span><br/>
 
                 <div className="text-center">
-                    <button onClick={this.ConnectWithKey.bind(this)} disabled={this.state.private_key.length == 0}>Connect Wallet</button>
+                    <button onClick={this.ConnectWithKey.bind(this)} disabled={private_key.length == 0}>Connect Wallet</button>
                 </div>
             </div>
         )
     }
 
-    
-    KeyConnect() {
-        const tabs = {
-            names: ["Encrypted Key", "Private Key"],
-            selectedTabIndex: this.state.selectedTabIndex,
-        }
-
+    AccountSelect() {
+        const { dispatch, mobile_nav } = this.props
+        const { accounts_list, private_key, storeId } = this.state
         return (
-            <div id="key-connect">
-                <div className="d-flex justify-content-between">
-                    <h4>CONNECT WALLET</h4>
-                    <div className="link mr-5" onClick={() => this.resetInputs({dialogType: "create"})}>Don’t have an account?</div>
+            <div className="input-container">
+                <h4>Select the account you want to use:</h4>
+                <div className="d-flex mt-3">
+                    <select id="mult-acc-select" className="w-100 mr-3" defaultValue={storeId}>
+                        {accounts_list.map(account => {
+                            return <option key={account.id} value={account.id}>{account.name}</option>
+                        })}
+                    </select>
+                    <button
+                        onClick={() => {
+                            const account = document.getElementById('mult-acc-select').value
+                            dispatch(ConnectAccount(account, private_key)).then(() => {
+                                if (mobile_nav) {
+                                    mobile_nav()
+                                }
+                            })
+                        }}
+                    >
+                        Continue
+                    </button>
                 </div>
-                <QTTabBar
-                    className="underline small static set-width qt-font-bold d-flex justify-content-left"
-                    width={120}
-                    gutter={10}
-                    tabs={tabs}
-                    switchTab={this.handleSwitch.bind(this)}
-                />
-                
-                {this.state.selectedTabIndex == 0 ? 
-                    (this.state.encryptStep == 0 ? <this.ConnectEncrypted /> : <this.EncryptKey />)
-                    : <this.ConnectPrivateKey />}
             </div>
         )
     }
 
     KeyCreate() {
+        const { isMobile, network } = this.props
+        const { regStep, referrer, referrer_error } = this.state
         return (
             <div id="key-create">
-                <div className="d-flex justify-content-between">
-                    <h4>CREATE WALLET</h4>
-                    <div className="link mr-5" onClick={() => this.resetInputs({dialogType: "connect"})}>Already have a key?</div>
-                </div>
                 <div className="input-container">
-                    <p className="info">
-                        The QUANTA blockchain is Graphene-based Architecture which uses 
-                        an account system based on username, and public-private key signature. 
-                        This wallet creation will generate you a random public-private key, 
-                        and register your account with the blockchain, then encrypt your private 
-                        key with a password into a private “json” key to download to your computer. 
-                        Beware, if you lose the password, you will lose your funds forever.
-                    </p>
-
-                    <div className="mb-2">
-                        <label>USERNAME</label><br/>
-                        <input id="name-input" type="text" autoComplete="off" placeholder="Username" spellCheck="false" 
-                            value={this.state.username} onChange={(e) => this.setState({username: e.target.value})}/>
-                    </div>
-                    
-
-                    <div className="mb-2">
-                        <label>PASSWORD</label><br/>
-                        <input id="pw-input" type="password" placeholder="Password"
-                            value={this.state.password} onChange={(e) => this.setState({password: e.target.value})}/>
-                    </div>
-
-                    <div className="mb-2">
-                        <label>CONFIRM PASSWORD</label><br/>
-                        <input id="pwconf-input" type="password" placeholder="Confirm Password" spellCheck="false" 
-                            value={this.state.confirm_password} onChange={(e) => this.setState({confirm_password: e.target.value})}/>
-                    </div>
-
-                    <span className="error" hidden={!this.state.authError}>{this.state.errorMsg}</span><br/>
-
-                    <div className="text-center">
-                        <button onClick={this.registerAccount.bind(this)} disabled={this.state.username.length == 0 || this.state.password.length == 0 || this.state.confirm_password.length == 0}>
-                            {this.state.processing ? <Loader /> : "REGISTER ACCOUNT"}
-                        </button>
-                    </div>
+                    {referrer ? 
+                        <div className="referral text-right small mb-1">
+                            <b>Referral:</b> {referrer}
+                            {referrer_error ? <span className="text-danger"><br/>{referrer_error}</span> : null}
+                        </div>
+                    : null
+                    }
+                    {this.keyCreateStep(regStep)}
                 </div>
             </div>
         )
     }
 
     KeyDownload() {
+        const { isMobile, network, mobile_nav } = this.props
+        const { authError, errorMsg, no_email, personal_key, downloaded } = this.state
         return (
-            <div id="key-create">
-                <div className="d-flex justify-content-between">
-                    <h4>CREATE WALLET</h4>
-                    <div className="link mr-5" onClick={() => this.setState({dialogType: "connect"})}>Already have a key?</div>
-                </div>
-                <div className="input-container">
-                    <h5>ACCOUNT REGISTERED</h5>
-                    <div className="warning qt-font-light">
-                        <h5>IMPORTANT INFORMATION</h5>
-                        <ul>
-                            <li>Store this wallet securely. QUANTA does not have your keys.</li>
-                            <li>If you lose it you will lose your tokens.</li>
-                            <li>Do not share it! Your funds will be stolen if you use this file on a malicious/phishing site.</li>
-                            <li>Make a backup! Secure it like the millions of dollars it may one day be worth.</li>
-                            <li>This is not a ERC-20.</li>
-                        </ul>
-                    </div>
-
-                    <div className="agreements my-5">
-                        <form name="agreements">
-                            <div>
-                                <input type="checkbox" id="pw" name="pw"/>
-                                <label htmlFor="pw">I have remembered or otherwise stored my password.</label>
-                            </div>
-                            <div>
-                                <input type="checkbox" id="rc" name="rc" />
-                                <label htmlFor="rc">I understand that no one can recover my password or file 
-                                    if I lose or forget it. Thus if I lose access to my account, I will lose 
-                                    access to my funds without a recovery opportunity.</label>
-                            </div>
-                            <div>
-                                <input type="checkbox" id="ts" name="ts" />
-                                <label htmlFor="ts">I agree with Terms of Service and Privacy Policy.</label>
-                            </div>
+            <React.Fragment>
+                <h5>ACCOUNT REGISTERED</h5>
+                <div className="warning qt-font-light">
+                    <h5>IMPORTANT INFORMATION</h5>
+                    <ul>
+                        <li>We do not store your encrypted or unencrypted key.</li>
+                        {no_email ?
+                            <li>You have elected not to use email backup, keep your private key safe.</li>
+                        :
+                            <li>We have emailed you your encrypted key and a QR code to login on your mobile app.</li>
+                        }
+                        {personal_key ?
+                            null
+                        :
+                            <li>Your password is used to decrypt the wallet. We cannot recover this if you lose it.</li>
+                        }
                         
-                        </form>
-                    </div>
-                    <span className="error" hidden={!this.state.authError}>{this.state.errorMsg}</span><br/>
-                    <div className="text-center">
-                        <button className="mb-2" onClick={this.DownloadKey}>DOWNLOAD FILE</button>
-                        <div className={"link qt-font-small" + (!this.state.downloaded ? " invisible" : "")} onClick={() => this.setState({dialogType: "connect"})}>
+                    </ul>
+                </div>
+
+                <div className="text-center mt-5">
+                    {window.isApp ?
+                        <button className="mb-2" onClick={mobile_nav}>CONTINUE</button>
+                    :
+                    personal_key ? 
+                        null
+                        :
+                        <button className="mb-2" onClick={this.downloadKey}>DOWNLOAD JSON</button>
+                    }
+
+                    { personal_key || downloaded ?
+                        <div className="link qt-font-small text-center mb-2" 
+                            onClick={() => this.resetInputs({dialogType: "connect"})}>
                             <u>Proceed to Connect your Wallet</u>
                         </div>
-                    </div>
-
+                        : null
+                    }
                 </div>
-            </div>
+            </React.Fragment>
         )
     }
 
     render() {
-
+        const { isMobile, network } = this.props
+        const { dialogType, selectedTabIndex, encryptStep, account_select } = this.state
         return (
-            <div id="connect-dialog" className={dialog + (this.props.isMobile ? " mobile" : "")} style={{display: "none"}} 
+            <div id="connect-dialog" className={dialog + " d-flex align-content-center qt-font-regular" + (isMobile ? " mobile" : "")} 
                 onDragOver={(e)=> e.preventDefault()} onDrop={(e) => e.preventDefault()}>
-                <div className="container">
-                    <div className="close-btn" onClick={() => this.closeDialog()}><img src={devicePath("public/images/close_btn.svg")} /></div>
-                    {this.state.dialogType == "create" ? 
-                        this.state.regStep == 1 ? <this.KeyCreate /> : <this.KeyDownload /> :
-                        <this.KeyConnect />
+                <div className={"container " + network}>
+                    {!isMobile ? 
+                        <div className="close-btn" onClick={this.closeDialog.bind(this)}><img src={devicePath("public/images/close_btn.svg")} /></div> 
+                        : null
                     }
-                    <p>Your private keys are not sent to QUANTA. All transactions are signed within your browser <br/>
+                    <h3>CONNECT {network == "testnet" ? "TESTNET" : ""} WALLET</h3>
+
+                    <div className="my-3">
+                        <div className={dialogType == "create" ? "active" : ""}>
+                            <div className="selection-tab d-flex"
+                                onClick={() => this.setState({dialogType: "create"})}
+                            >
+                                <div className="select-btn"><div></div></div>
+                                <b>Create a Wallet.</b>&nbsp;New to Quanta?
+                            </div>
+                            { dialogType == "create" ? <this.KeyCreate /> : null }
+                        </div>
+
+                        <div className={"border-top border-bottom " + (dialogType == "connect" && selectedTabIndex == 0 ? "active" : "")}
+                            onClick={() => this.setState({dialogType: "connect", selectedTabIndex: 0})}
+                        >
+                            <div className="selection-tab d-flex">
+                                <div className="select-btn"><div></div></div>
+                                <b>Connect with Encrypted Key</b>
+                            </div>
+                            { dialogType == "connect" && selectedTabIndex == 0 ? 
+                                account_select ? <this.AccountSelect /> : encryptStep == 0 ? <this.ConnectEncrypted /> : <this.EncryptKey />
+                                : null 
+                            }
+                        </div>
+
+                        <div className={dialogType == "connect" && selectedTabIndex == 1 ? "active" : ""}
+                            onClick={() => this.setState({dialogType: "connect", selectedTabIndex: 1})}
+                        >
+                            <div className="selection-tab d-flex">
+                                <div className="select-btn"><div></div></div>
+                                <b>Connect with Private Key</b>
+                            </div>
+                            { dialogType == "connect" && selectedTabIndex == 1 ? 
+                                account_select ? <this.AccountSelect /> : <this.ConnectPrivateKey />
+                                : null 
+                            }
+                        </div>
+                    </div>
+
+                    <p className="qt-font-extra-small qt-white-62 qt-font-light m-0">Your private keys are not sent to QUANTA. All transactions are signed within your browser 
                         and keys are not exposed over the internet.</p>
                 </div>
             </div>
@@ -686,7 +1276,9 @@ class ConnectDialog extends Component {
 } 
 
 const mapStateToProps = (state) => ({
-    currentTicker: state.app.currentTicker
+    network: state.app.network,
+    private_key: state.app.private_key,
+    isMobile: state.app.isMobile
 });
-export default connect(mapStateToProps)(ConnectDialog)
-export { Connect, ConnectLink }
+
+export default connect(mapStateToProps)(Connect)
