@@ -3,7 +3,6 @@ import Header from './headersimple.jsx';
 import { connect } from 'react-redux'
 import { css } from 'emotion'
 import globalcss from './global-css.js'
-import MobileHeader from './ui/mobileHeader.jsx';
 import QTTabBar from './ui/tabBar.jsx'
 import { PrivateKey, PublicKey, Signature } from "@quantadex/bitsharesjs";
 
@@ -51,6 +50,11 @@ const container = css`
         button:hover {
             box-shadow: 0 0 5px #777;
         }
+        button:disabled, button:disabled:hover {
+            background: #444;
+            color: #999;
+            box-shadow: none;
+        }
     }
 
     .verify {
@@ -64,6 +68,8 @@ const container = css`
     }
 
     &.mobile {
+        background-color: transparent;
+        min-height: unset;
         padding: 0;
         .row {
             margin: 0;
@@ -138,7 +144,9 @@ class Message extends Component {
                 <div>
                     <textarea id='message' className="d-block" autoFocus placeholder="Your message here..."></textarea>
                 </div>
-                <button className="my-4 cursor-pointer" onClick={this.signMessage}>Sign</button>
+                <button disabled={!this.props.private_key} className="my-4 cursor-pointer" onClick={this.signMessage}>
+                    {this.props.private_key ? "Sign" : "LOCK"}
+                </button>
                 <div>
                     <textarea className="d-block" readOnly onClick={(e) => this.selectText(e)} value={this.state.signedMsg} >
                     </textarea>
@@ -169,11 +177,11 @@ class Message extends Component {
         
 
     render() {
-        if (this.props.private_key == null) {
-            window.location.assign('/exchange')
-            return
-        } 
-
+        // if (publicKey == null) {
+        //     window.location.assign('/exchange')
+        //     return
+        // } 
+        const { isMobile } = this.props
         const tabs = {
             names: ['Sign Message', 'Verify Message'],
             selectedTabIndex: 0,
@@ -182,9 +190,9 @@ class Message extends Component {
         const content = [<this.Sign />, <this.Verify />]
 
         return (
-            <div className={container + " container-fluid" + (this.props.isMobile ? " mobile" : "")}>
-                {this.props.isMobile ? 
-                    <MobileHeader />
+            <div className={container + " container-fluid" + (isMobile ? " mobile" : "")}>
+                {isMobile ? 
+                    null
                     :
                     <div className="row header-row">
                     <Header />
@@ -193,8 +201,8 @@ class Message extends Component {
                 <div className="tab-row row d-flex flex-column align-items-center">
                     <div className="tabs">
                         <QTTabBar
-                            className="underline static set-width qt-font-bold d-flex justify-content-left"
-                            width={130}
+                            className="underline static even-width qt-font-bold d-flex justify-content-left"
+                            width={isMobile || 130}
                             gutter={10}
                             tabs = {tabs}
                             switchTab = {this.handleSwitch.bind(this)}
