@@ -1,17 +1,16 @@
 import React, { Component } from 'react';
 import { css } from 'emotion'
 import { connect } from 'react-redux'
-import { UPDATE_NETWORK, LOGIN, switchTicker } from '../redux/actions/app.jsx'
 
 const container = css`
     display: flex;
-    bottom: 0;
     width: 100%;
     white-space: nowrap;
     line-height: 25px;
     z-index: 2;
+    user-select: none;
 
-    .blocknum, .avg-lat, .net-select {
+    .blocknum, .avg-lat, .net-select, .support {
         border-left: 1px solid #444;
         padding-right: 10px;
     }
@@ -23,6 +22,12 @@ const container = css`
     .blocknum {
         background: url(${devicePath("public/images/blockchain.svg")}) no-repeat 10px;
         padding-left: 40px;
+    }
+    .support {
+        padding-left: 10px;
+        a {
+            color: #1cdad8 !important;
+        }
     }
     .avg-lat {
         background: url(${devicePath("public/images/time.svg")}) no-repeat 10px;
@@ -67,6 +72,11 @@ const container = css`
             display: none;
         }
 
+        .support {
+            width: 100%;
+            text-align: center;
+        }
+
         .explorer, .avg-lat {
             display: inline-block;
         }
@@ -77,38 +87,26 @@ const container = css`
 `;
 
 class Status extends Component {
-    constructor(props) {
-        super(props);
-        this.state= {
-            network: this.props.network
-        }
-    }
-    switchExNet(network) {
-        if (network == this.state.network) return
-        const defaultTicker = network == "MAINNET" ? 'QDEX/ETH' : 'ETH/USD'
-        this.setState({network})
-        this.props.dispatch({
-            type: LOGIN,
-            private_key: null
-        });
-        this.props.dispatch({
-			type: UPDATE_NETWORK,
-			data: network
-        });
+    switchExNet(net) {
+        let current = window.location.pathname.startsWith("/testnet") ? "testnet" : "mainnet"
+        if (net == current) return
+        const defaultTicker = net == "mainnet" ? 'ETH_BTC' : 'ETH_USD'
 
-        this.props.dispatch(switchTicker(defaultTicker, true));
+        window.location = "/" + net + "/exchange/" + defaultTicker
     }
     
     render() {
-        const netsList = [{name: "MAINNET"}, {name: "TESTNET"}]
+        const { mobile, blockInfo } = this.props
+        const netsList = [{name: "mainnet"}, {name: "testnet"}]
 
         return (
-            <div id="quanta-status" className={container + (this.props.mobile ? " mobile" : "")}>
+            <div id="quanta-status" className={container + (mobile ? " mobile" : "")}>
                 <div className="brand">QUANTA Fair Trading Protocol</div>
-                <div className="blocknum" title="Highest Block">{this.props.blockInfo.blockNumber}</div>
-                <div className="avg-lat" title="Average Block Latency">{ this.props.blockInfo.blockTime}ms</div>
-                <div className="net-select position-relative">
-                    {this.state.network}
+                <div className="support"><a href="https://quantadex.zendesk.com/hc/en-us" target="_blank">{mobile ? "Help" : "Customer Support"}</a></div>
+                <div className="blocknum" title="Highest Block">{blockInfo.blockNumber}</div>
+                <div className="avg-lat" title="Average Block Latency">{blockInfo.blockTime}ms</div>
+                <div className="net-select position-relative text-uppercase">
+                    {window.location.pathname.startsWith("/testnet") ? "testnet" : "mainnet"}
                     <div className="net-options">
                         {netsList.map(item => {
                             return <div key={item.name} onClick={() => this.switchExNet(item.name)}>{item.name}</div>
