@@ -1,21 +1,10 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux'
-import moment from 'moment';
-
 import { css } from 'emotion'
-import { toast } from 'react-toastify';
-import globalcss from './global-css.js'
-
-import QTTabBar from './ui/tabBar.jsx'
-import QTDropdown from './ui/dropdown.jsx'
-import QTButton from './ui/button.jsx'
-import { Token, SmallToken } from './ui/ticker.jsx'
+import { SmallToken } from './ui/ticker.jsx'
 import Loader from './ui/loader.jsx'
 import Utils from '../common/utils'
-import lodash from 'lodash';
-
-import { buyTransaction } from "../redux/actions/app.jsx";
-import { sellTransaction } from "../redux/actions/app.jsx";
+import { buyTransaction, sellTransaction, TOGGLE_CONNECT_DIALOG } from "../redux/actions/app.jsx";
 import ReactGA from 'react-ga';
 
 const container = css`
@@ -63,6 +52,13 @@ const container = css`
     }
   }
 
+  button.connect-btn {
+    background-color: transparent;
+    border: 2px solid #66d7d7;
+    border-radius: 5px;
+    color: #66d7d7;
+  }
+
   .buy-btn, .sell-btn {
     height: 37px;
   }
@@ -74,7 +70,15 @@ const container = css`
   }
 
   .input-container {
-    margin: 15px 0;
+    display: flex;
+    align-items: center;
+    margin: 6px 0;
+
+    label {
+      flex: 1 1 60px;
+      margin: 0;
+    }
+
     input {
       border-radius: 2px;
       padding: 0 40px 0 10px;
@@ -87,6 +91,18 @@ const container = css`
     input:invalid {
       box-shadow: none;
     }
+    
+    .percent-container, input {
+      flex: 1 1 calc(100% - 60px);
+    }
+
+    .amount-select {
+      line-height: 17px; 
+      width: 23%;
+      border: 1px solid rgba(255,255,255,0.27);
+      border-radius: 2px;
+    }
+
     span {
       position: absolute;
       right: 10px;
@@ -210,7 +226,7 @@ class Trade extends Component {
     constructor(props) {
         super(props);
         this.state = {
-            trade_side: 0,
+            trade_side: this.props.trade_side || 0,
             processing: false,
             qty: 0,
             price: 0,
@@ -223,29 +239,31 @@ class Trade extends Component {
     }
 
     componentDidMount() {
-      if (!this.props.mobile) {
+      const { currentTicker, mobile, inputSetTime, currentPrice, inputBuyAmount, inputBuy} = this.props
+      if (!mobile) {
         return
       }
-      if (this.props.inputSetTime && this.props.currentPrice.price && (new Date() - this.props.inputSetTime) < 50) {
+      if (inputSetTime && currentPrice.price && (new Date() - inputSetTime) < 50) {
         this.setState({
-          qty: parseFloat(this.props.inputBuyAmount),
-          price: parseFloat(this.props.inputBuy),
-          total: parseFloat(this.props.inputBuyAmount) * parseFloat(this.props.inputBuy),
-          inputSetTime: this.props.inputSetTime,
-          currentPrice: this.props.currentPrice.price,
-          currentTicker: this.props.currentTicker
+          qty: parseFloat(inputBuyAmount),
+          price: parseFloat(inputBuy),
+          total: parseFloat(inputBuyAmount) * parseFloat(inputBuy),
+          inputSetTime: inputSetTime,
+          currentPrice: currentPrice.price,
+          currentTicker: currentTicker
         })
       } else {
         this.setState({
-          price: parseFloat(this.props.currentPrice.price),
-          currentPrice: this.props.currentPrice.price,
-          currentTicker: this.props.currentTicker
+          price: parseFloat(currentPrice.price),
+          currentPrice: currentPrice.price,
+          currentTicker: currentTicker
         })
       }
     }
 
     componentWillReceiveProps(nextProps) {
-        if (!this.props.mobile && nextProps.inputSetTime != undefined && nextProps.inputSetTime != this.state.inputSetTime) {
+        const { inputSetTime, currentPrice, currentTicker } = this.state
+        if (nextProps.inputSetTime != undefined && nextProps.inputSetTime != inputSetTime) {
             this.setState({
                 qty: parseFloat(nextProps.inputBuyAmount),
                 price: parseFloat(nextProps.inputBuy),
@@ -253,82 +271,45 @@ class Trade extends Component {
                 inputSetTime: nextProps.inputSetTime
             })
         }
-        if (this.state.currentPrice !== nextProps.currentPrice.price && (this.state.currentTicker !== nextProps.currentPrice.ticker || !this.state.currentPrice)) {
+        if (currentPrice !== nextProps.currentPrice.price && (currentTicker !== nextProps.currentPrice.ticker || !currentPrice)) {
           this.setState({
             currentTicker: nextProps.currentPrice.ticker, 
             currentPrice: nextProps.currentPrice.price, 
             price: nextProps.currentPrice.price, 
-            qty: 0
+            qty: 0,
+            total: 0
           })
         }
     }
 
-    notify_success = (toastId, msg) => toast.update(toastId, {
-        render: msg,
-        type: toast.TYPE.SUCCESS,
-        autoClose: 5000,
-        position: toast.POSITION.TOP_CENTER
-    });
-    notify_failed = (toastId, msg) => toast.update(toastId, {
-        render: msg,
-        type: toast.TYPE.ERROR,
-        autoClose: 5000,
-        position: toast.POSITION.TOP_CENTER
-    });
-
-    toastMsg(label, success, e) {
-        const msg = (<div>
-            <span>{label}</span><br />
-            <span>{success ? "Order ID: " + e.trx.operation_results[0][1] :
-                "Failed order: " + (e.message.includes("insufficient balance") ? "Insufficient Balance" : "Unable to place order")}</span>
-        </div>)
-        return msg
-    }
-
     handleBuy(e) {
-        const label = this.props.currentTicker.split('/')[1] + " " + this.state.price + " @ " + this.state.qty
-        const toastId = toast("BUYING " + label, { autoClose: false, position: toast.POSITION.TOP_CENTER });
+        const { dispatch, currentTicker, mobile_nav } = this.props
+        const { price, qty } = this.state
 
         ReactGA.event({
             category: 'BUY',
-            action: this.props.currentTicker
+            action: currentTicker
         });
 
         this.setState({ processing: true })
-        let price = Utils.maxPrecision(this.state.price, window.assetsBySymbol[this.props.currentTicker.split('/')[1]].precision)
-        let amount = Utils.maxPrecision(this.state.qty, window.assetsBySymbol[this.props.currentTicker.split('/')[0]].precision)
-        this.props.dispatch(buyTransaction(this.props.currentTicker, price, amount))
-            .then((e) => {
-                const msg = this.toastMsg("BUY " + label, true, e)
-                this.notify_success(toastId, msg)
-            }).catch((e) => {
-                const msg = this.toastMsg("BUY " + label, false, e)
-                this.notify_failed(toastId, msg)
-            }).finally(() => {
+        dispatch(buyTransaction(currentTicker, price, qty, false, mobile_nav))
+            .finally(() => {
                 this.setState({ processing: false })
             })
     }
 
     handleSell(e) {
-        const label = this.props.currentTicker.split('/')[1] + " " + this.state.price + " @ " + this.state.qty
-        const toastId = toast("SELLING " + label, { autoClose: false, position: toast.POSITION.TOP_CENTER });
+        const { dispatch, currentTicker, mobile_nav } = this.props
+        const { price, qty } = this.state
 
         ReactGA.event({
             category: 'SELL',
-            action: this.props.currentTicker
+            action: currentTicker
         });
 
         this.setState({ processing: true })
-        let price = Utils.maxPrecision(this.state.price, window.assetsBySymbol[this.props.currentTicker.split('/')[1]].precision)
-        let amount = Utils.maxPrecision(this.state.qty, window.assetsBySymbol[this.props.currentTicker.split('/')[0]].precision)
-        this.props.dispatch(sellTransaction(this.props.currentTicker, price, amount))
-            .then((e) => {
-                const msg = this.toastMsg("SELL " + label, true, e)
-                this.notify_success(toastId, msg)
-            }).catch((e) => {
-                const msg = this.toastMsg("SELL " + label, false, e)
-                this.notify_failed(toastId, msg)
-            }).finally(() => {
+        dispatch(sellTransaction(currentTicker, price, qty, mobile_nav))
+            .finally(() => {
                 this.setState({ processing: false })
             })
     }
@@ -338,53 +319,58 @@ class Trade extends Component {
     }
 
     setPercentAmount(perc, symbol) {
-      const coin = window.assetsBySymbol[symbol].id
-      const amount = this.props.balance[coin].balance * (parseInt(perc)/100)
+      const { balance } = this.props
+      const { trade_side, price} = this.state
+      const amount = balance[symbol] ? balance[symbol].balance * (parseInt(perc)/100) : 0
 
       let qty, total
-      if (this.state.trade_side == 0) {
-        qty = amount / this.state.price
+      if (trade_side == 0) {
+        qty = amount / price
         total = amount
       } else {
         qty = amount
-        total = qty * this.state.price
+        total = qty * price
       }
       this.setState({qty, total})
     }
 
     render() {
-        // const tabs = {
-        //     names: ['LIMIT', 'MARKET', 'STOP-LIMIT'],
-        //     selectedTabIndex: 0,
-        // }
-
+        const { currentTicker, balance, mobile, mobile_nav, fee, publicKey, private_key } = this.props
+        const { trade_side, price, qty, total, processing} = this.state
         const dropdown_items = {
             items: ["25%", "50%", "75%", "100%"],
             value: "25%"
         }
+        
+        if (currentTicker == null) {
+          return <div></div>;
+        }
 
-        const tradingPair = this.props.currentTicker.split('/')
-        const precisions = window.assetsBySymbol && [window.assetsBySymbol[tradingPair[0]] ? window.assetsBySymbol[tradingPair[0]].precision : 5, window.assetsBySymbol[tradingPair[1]] ? window.assetsBySymbol[tradingPair[1]].precision : 5] || [0,0]
-        const balance = this.props.balance
+        const tradingPair = currentTicker.split('/')
+        const base = window.assetsBySymbol && window.assetsBySymbol[tradingPair[0]]
+        const counter = window.assetsBySymbol && window.assetsBySymbol[tradingPair[1]]
+        const precisions = [base ? base.precision : 5, counter ? counter.precision : 5] || [0,0]
         var pairBalance = []
-        Object.keys(balance).forEach((currency) => {
-            if (!window.assets[balance[currency].asset]) {
+        Object.keys(balance).forEach((symbol) => {
+            if (!window.assetsBySymbol[symbol]) {
               return
             }
             var item = {}
-            item.currency = window.assets[balance[currency].asset].symbol
+            item.currency = symbol
             if (!tradingPair.includes(item.currency)) {
                 return
             }
-            item.amount = balance[currency].balance ? balance[currency].balance : 0
+            item.amount = balance[symbol].balance ? balance[symbol].balance : 0
             pairBalance.push(item)
         })
-        
+        const taker_fee = (window.assetsBySymbol && window.assetsBySymbol[tradingPair[trade_side]].options.market_fee_percent/100) || 0
+        const maker_fee = (window.maker_rebate_percent_of_fee && taker_fee - (taker_fee * (window.maker_rebate_percent_of_fee/10000))) || 0
+
         return (
-            <div className={container + " container-fluid" + (this.props.mobile ? " mobile" : "")}>
+            <div className={container + " container-fluid" + (mobile ? " mobile" : "")}>
                 <div className="buy-sell-toggle">
-                    <button id="buy-switch" className={"buy-btn" + (this.state.trade_side !== 0 ? " inactive" : "")} onClick={this.switchTradeTo.bind(this, 0)}>BUY {tradingPair[0]}</button>
-                    <button id="sell-switch" className={"sell-btn" + (this.state.trade_side !== 1 ? " inactive" : "")} onClick={this.switchTradeTo.bind(this, 1)}>SELL {tradingPair[0]}</button>
+                    <button id="buy-switch" className={"buy-btn" + (trade_side !== 0 ? " inactive" : "")} onClick={this.switchTradeTo.bind(this, 0)}>BUY <SmallToken name={tradingPair[0]} /></button>
+                    <button id="sell-switch" className={"sell-btn" + (trade_side !== 1 ? " inactive" : "")} onClick={this.switchTradeTo.bind(this, 1)}>SELL <SmallToken name={tradingPair[0]} /></button>
                 </div>
 
                 <div className="transac-actions">
@@ -393,43 +379,46 @@ class Trade extends Component {
                         <input type="number" className="trade-input qt-number-bold qt-font-small" title=""
                             name="price"
                             autoComplete="off"
-                            onFocus={(e) => e.target.select()}
+                            onFocus={(e) => !mobile && e.target.select()}
                             min="0"
-                            value={Utils.maxPrecision(this.state.price, precisions[1])}
+                            value={price == "-" ? 0 : price}
                             onChange={(e) => {
-                              let value = Utils.maxPrecision(e.target.value, precisions[1])
+                              let value = e.target.value
                               this.setState({
                               price: value,
-                              total: this.state.qty * value
+                              total: qty * value
                             })}} />
                         <SmallToken name={tradingPair[1]} />
                     </div>
                     <div className="input-container">
-                        <label>AMOUNT</label>
-                        <div className="d-flex">
-                          <QTDropdown
-                            items={dropdown_items.items}
-                            value={dropdown_items.value}
-                            className="down bordered dark qt-font-base qt-font-bold"
-                            reverse={true}
-                            width="58"
-                            height="32"
-                            onChange={(e) => this.props.balance && this.setPercentAmount(e, tradingPair[this.state.trade_side == 0 ? 1 : 0])}/>
-                          <input type="number" className="trade-input qt-number-bold qt-font-small rounded-0-left" title=""
-                              name="amount"
-                              autoComplete="off"
-                              onFocus={(e) => e.target.select()}
-                              min="0"
-                              value={Utils.maxPrecision(this.state.qty, precisions[0])}
-                              onChange={(e) => {
-                                let value = Utils.maxPrecision(e.target.value, precisions[0])
-                                this.setState({
-                                qty: value,
-                                total: value * this.state.price
-                              })}} />
-                          <SmallToken name={tradingPair[0]} />
+                      <label>AMOUNT</label>
+                        <input type="number" className="trade-input qt-number-bold qt-font-small rounded-0-left" title=""
+                            name="amount"
+                            autoComplete="off"
+                            onFocus={(e) => !mobile && e.target.select()}
+                            min="0"
+                            value={Utils.maxPrecision(qty, precisions[0])}
+                            onChange={(e) => {
+                              let value = Utils.maxPrecision(e.target.value, precisions[0])
+                              this.setState({
+                              qty: value,
+                              total: value * price
+                            })}} />
+                        <SmallToken name={tradingPair[0]} />
+                    </div>
+                    <div className="input-container">
+                        <label className="invisible">Percent</label>
+                        <div className="percent-container d-flex justify-content-between text-center">
+                          {dropdown_items.items.map(item => {
+                            return (
+                              <div key={item} className="amount-select cursor-pointer"
+                                onClick={() => balance && this.setPercentAmount(item, tradingPair[trade_side == 0 ? 1 : 0])}
+                              >
+                                {item}
+                              </div>
+                            )
+                          })}
                         </div>
-                        
                     </div>
                     <div className="input-container">
                         <label>TOTAL</label>
@@ -437,19 +426,23 @@ class Trade extends Component {
                             title=""
                             type="number"
                             autoComplete="off"
-                            onFocus={(e) => e.target.select()}
+                            onFocus={(e) => !mobile && e.target.select()}
                             className="trade-input qt-number-bold qt-font-small"
                             min="0"
-                            value={Utils.maxPrecision(this.state.total, precisions[1])}
+                            value={total < 1/Math.pow(10, precisions[1]) ? total : Utils.maxPrecision(total, precisions[1])}
                             onChange={(e) => {
                               let value = Utils.maxPrecision(e.target.value, precisions[1])
                               this.setState({
-                              qty: value/this.state.price,
+                              qty: value/price,
                               total: value
                             })}}
                             />
                         <SmallToken name={tradingPair[1]} />
                     </div>
+                    {total < 1/Math.pow(10, precisions[1]) && !(price <= 0 || qty <= 0) ? 
+                      <div className="text-danger text-right">* total must be more than {1/Math.pow(10, precisions[1])}</div> 
+                      : null
+                    }
 
                     <div className="fees-container">
                         Estimate Fees
@@ -457,44 +450,76 @@ class Trade extends Component {
                             <tbody>
                                 <tr>
                                     <td>Maker</td>
-                                    <td className="text-left text-muted pl-3">{window.assetsBySymbol && (window.assetsBySymbol[tradingPair[this.state.trade_side]].options.market_fee_percent)/100}%</td>
+                                    <td className="text-left text-muted pl-3">{maker_fee}%</td>
                                     <td className="text-right pr-2">{window.assetsBySymbol 
-                                        && window.assetsBySymbol[tradingPair[this.state.trade_side]].options.market_fee_percent != 0 
-                                        && ((((this.state.qty * Math.pow(10, 6)) * (this.state.price * Math.pow(10, 6))) / Math.pow(10, 12))*((window.assetsBySymbol[tradingPair[this.state.trade_side]].options.market_fee_percent)/10000)).toLocaleString(navigator.language, {maximumFractionDigits: 6}) || 0}</td>
-                                    <td className="text-muted">{tradingPair[this.state.trade_side]}</td>
+                                        && window.assetsBySymbol[tradingPair[trade_side]].options.market_fee_percent != 0 
+                                        && ((trade_side == 0 ? qty : total)*maker_fee/100).toLocaleString(navigator.language, {maximumFractionDigits: precisions[trade_side]}) || 0}</td>
+                                    <td className="text-muted"><SmallToken name={tradingPair[trade_side]} /></td>
                                 </tr>
                                 <tr>
                                     <td>Taker</td>
-                                    <td className="text-left text-muted pl-3">{window.assetsBySymbol && window.assetsBySymbol[tradingPair[this.state.trade_side]].options.market_fee_percent}%</td>
+                                    <td className="text-left text-muted pl-3">{taker_fee}%</td>
                                     <td className="text-right pr-2">{window.assetsBySymbol 
-                                        && window.assetsBySymbol[tradingPair[this.state.trade_side]].options.market_fee_percent != 0 
-                                        && ((((this.state.qty * Math.pow(10, 6)) * (this.state.price * Math.pow(10, 6))) / Math.pow(10, 12))*((window.assetsBySymbol[tradingPair[this.state.trade_side]].options.market_fee_percent)/10000).toLocaleString(navigator.language, {maximumFractionDigits: 6})) || 0}</td>
-                                    <td className="text-muted">{tradingPair[this.state.trade_side]}</td>
+                                        && window.assetsBySymbol[tradingPair[trade_side]].options.market_fee_percent != 0 
+                                        && ((trade_side == 0 ? qty : total)*taker_fee/100).toLocaleString(navigator.language, {maximumFractionDigits: precisions[trade_side]}) || 0}</td>
+                                    <td className="text-muted"><SmallToken name={tradingPair[trade_side]} /></td>
                                 </tr>
                                 <tr>
                                     <td colSpan="2">Platform Fees</td>
-                                    <td className="text-right pr-2">{this.props.fee.amount}</td>
-                                    <td className="text-muted">{this.props.fee.symbol}</td>
+                                    <td className="text-right pr-2">{fee.amount}</td>
+                                    <td className="text-muted"><SmallToken name={fee.symbol} /></td>
                                 </tr>
                             </tbody>
                         </table>
                     </div>
 
-                    { this.props.private_key ? 
-                        <div>
-                            {this.state.trade_side == 1 ?
-                                <button id="sell-action" className="sell-btn" disabled={this.state.price <= 0 || this.state.qty <= 0 || this.state.processing}
-                                    onClick={this.handleSell.bind(this)}>
-                                    {this.state.processing ? <Loader /> : "PLACE SELL ORDER"}
-                                </button>
-                                :
-                                <button id="buy-action" className="buy-btn" disabled={this.state.price <= 0 || this.state.qty <= 0 || this.state.processing}
-                                    onClick={this.handleBuy.bind(this)}>
-                                    {this.state.processing ? <Loader /> : "PLACE BUY ORDER"}
-                                </button>
+                    { private_key ? 
+                      <div>
+                        {trade_side == 1 ?
+                          <button id="sell-action" className="sell-btn" 
+                            disabled={price <= 0 || qty <= 0 || total < 1/Math.pow(10, precisions[1]) || processing}
+                            onClick={this.handleSell.bind(this)}>
+                            {processing ? <Loader /> : 
+                              <span>
+                                PLACE SELL ORDER
+                              </span>
                             }
-                        </div>
-                    : "" }
+                          </button>
+                          :
+                          <button id="buy-action" className="buy-btn" 
+                            disabled={price <= 0 || qty <= 0 || total < 1/Math.pow(10, precisions[1]) || processing}
+                            onClick={this.handleBuy.bind(this)}>
+                            {processing ? <Loader /> : 
+                              <span>
+                                PLACE BUY ORDER
+                              </span>
+                            }
+                          </button>
+                        }
+                      </div>
+
+                      : 
+                      mobile || publicKey ? 
+                        <button className="connect-btn" 
+                        onClick={mobile && mobile_nav ? () => mobile_nav("connect") 
+                          : 
+                          () => this.props.dispatch({
+                            type: TOGGLE_CONNECT_DIALOG,
+                            data: "connect"
+                          })
+                        }>
+                        CONNECT WALLET TO TRADE
+                      </button> 
+                      : null
+                    }
+
+                    { mobile && publicKey ?
+                      <div className="d-flex justify-content-around flex-wrap qt-font-light qt-font-small text-secondary">
+                        <span className="mx-2">{base.symbol.split("0X")[0]} Balance: {balance[base.symbol] ? balance[base.symbol].balance : 0}</span>
+                        <span className="mx-2">{counter.symbol.split("0X")[0]} Balance: {balance[counter.symbol] ? balance[counter.symbol].balance : 0}</span>
+                      </div>
+                      : null
+                    }
                 </div>
             </div>
         )
@@ -503,13 +528,14 @@ class Trade extends Component {
 
 const mapStateToProps = (state) => ({
     private_key: state.app.private_key,
+    publicKey: state.app.publicKey,
     requestedPrice: 0,
     requestedQty: 0,
     bids: state.app.tradeBook.bids,
     asks: state.app.tradeBook.asks,
     currentTicker: state.app.currentTicker,
     currentPrice: state.app.mostRecentTrade,
-    balance: state.app.balance && lodash.keyBy(state.app.balance, "asset"),
+    balance: state.app.balance || {},
     inputBuy: state.app.inputBuy,
     inputBuyAmount: state.app.inputBuyAmount,
     inputSetTime: state.app.setTime,
