@@ -13,7 +13,7 @@ const container = css `
 export default class SearchBox extends Component {
     render() {
         return (
-            <input className={container} spellCheck="false" onClick={(e) => e.stopPropagation()}
+            <input className={container} spellCheck="false" autoComplete="off" onClick={(e) => e.stopPropagation()}
 				onChange={this.props.onChange} placeholder={this.props.placeholder} style={this.props.style} />
         )
     }
