@@ -3,9 +3,11 @@ import CONFIG from '../../config.js'
 import styled, { css } from 'react-emotion'
 import globalcss from '../global-css'
 import QTButton from './button.jsx'
+import Loader from './loader.jsx'
 
 const container = css`
   width:100%;
+  position: relative;
 
   thead {
     pointer-events: none;
@@ -42,6 +44,14 @@ const container = css`
 
   th.right, td.right {
     text-align: right;
+  }
+
+  th.left-mirror, td.left-mirror {
+    padding-left: 5px;
+  }
+
+  th.right-mirror, td.right-mirror {
+    padding-right: 5px;
   }
 
   th.center, td.center {
@@ -94,6 +104,22 @@ const container = css`
     right: 5px;
   }
 
+  .user-order::after {
+    content: "";
+    border: solid 4px transparent;
+    position: absolute;
+    transform: translateY(6px);
+  }
+
+  .left.user-order::after {
+    border-left-color: #fff;
+    left: 0;
+  }
+
+  .right.user-order::after {
+    border-right-color: #fff;
+    right: 0;
+  }
 `
 
 export default class QTTableViewSimple extends React.Component {
@@ -115,23 +141,25 @@ export default class QTTableViewSimple extends React.Component {
   }
 
   render() {
+    const { HideHeader, columns, ticker, dataSource, onAction, dispatch, disabled, cancelOrder,
+      max, barDir, barColor, cancelling, openConnect } = this.props
     return (
       <table className={container}>
 
-        <thead className={this.props.HideHeader ? "hidden" : ""}><tr>
+        <thead className={HideHeader ? "hidden" : ""}><tr>
           {
-            this.props.columns.map((col, index) => {
-              return <th key={index} className={col.float}>{typeof(col.name) === "string" ? col.name : col.name(this.props.ticker)}</th>
+            columns.map((col, index) => {
+              return <th key={index} className={col.float}>{typeof(col.name) === "string" ? col.name : col.name(ticker)}</th>
             })
           }
         </tr></thead>
         <tbody>
         {
-          this.props.dataSource.map((row, index) => {
+          dataSource.map((row, index) => {
             return (
-              <tr key={row.id ? row.id + index : index}
-                onClick={this.props.onAction ? () => this.props.onAction(row) : null}>
-                {this.props.columns.map((col, colindex) => {
+              <tr key={row.id ? row.id + index : index} className={row.is_user ? "user-order " + barDir : ""}
+                onClick={onAction ? () => onAction(row) : null}>
+                {columns.map((col, colindex) => {
                     var col_color = row[col.key];
                     if (row.color_key != undefined) {
                       col_color = row.color_key;
@@ -143,7 +171,7 @@ export default class QTTableViewSimple extends React.Component {
                           width="10"
                           height="10"
                           onClick={() => {
-                            this.props.dispatch(col.handleClick(row.pair))
+                            dispatch(col.handleClick(row.pair))
                           }} />
                         </td>
                       )
@@ -151,26 +179,29 @@ export default class QTTableViewSimple extends React.Component {
                     if (col.type == "id") {
                       return (
                         <td key={index + "-" + colindex} className={[col.float,col.color(col_color),col.fontWeight,col.fontSize].join(" ")}>
-                          <a href={CONFIG.SETTINGS[window.currentNetwork].EXPLORER_URL + "/object/" + row[col.key]} target="_blank">{row[col.key]}</a>
+                          <a href={CONFIG.getEnv().EXPLORER_URL + "/object/" + row[col.key]} target="_blank">{row[col.key]}</a>
                         </td>
                       )
                     }
                     if (col.type == "block") {
                       return (
                         <td key={index + "-" + colindex} className={[col.float,col.color(col_color),col.fontWeight,col.fontSize].join(" ")}>
-                          <a href={CONFIG.SETTINGS[window.currentNetwork].EXPLORER_URL + "/ledgers/" + row[col.key]} target="_blank">{row[col.key]}</a>
+                          <a href={CONFIG.getEnv().EXPLORER_URL + "/ledgers/" + row[col.key]} target="_blank">{row[col.key]}</a>
                         </td>
                       )
                     }
                     if (col.type == "cancel") {
                       return (
                         <td key={index + "-" + colindex} id={"cancel-" + row.id.replace(/\./g, '-')} className={col.float}>
-                            <div className="loader"></div>
+                          {cancelling && cancelling[row.id] ?
+                            <Loader size={15} margin="0 30px 0 auto" /> 
+                            :
                             <QTButton className="grey inverse qt-font-semibold qt-font-base" 
                             borderWidth="1" width="66" height="18" label="CANCEL"
                             onClick={() => {
-                              this.props.cancelOrder(row.assets, row.id)
+                              disabled ? openConnect() : cancelOrder(row.assets, row.id)
                             }}/>
+                          }
                         </td>
                         
                       )
@@ -185,10 +216,10 @@ export default class QTTableViewSimple extends React.Component {
                   })
                 }
                 
-                {this.props.max ? 
-                  <td className={"perc-bar " + this.props.barDir}
-                  style={{background: `rgba(${row.color_key != undefined ? (row.color_key == 0 ? "33,224,219" : "255,35,116") : this.props.barColor},${0.05 + Math.round((row.total / this.props.max)*100)/1000})`, 
-                          width: `${Math.round((row.total / this.props.max)*100)}%`}}></td> 
+                {max ? 
+                  <td className={"perc-bar " + barDir}
+                  style={{background: `rgba(${row.color_key != undefined ? (row.color_key == 0 ? "33,224,219" : "255,35,116") : barColor},${0.05 + Math.round((row.total / max)*100)/1000})`, 
+                          width: `${Math.round((row.total / max)*100)}%`}}></td> 
                 : null}
               </tr>
             )
