@@ -66,40 +66,63 @@ export default class TxDialog extends Component {
     }
 
     render() {
-        // console.log(this.props)
+        const {data} = this.props
+        const has_fund = data.fund ? data.fund > data.fee.amount : true
         return (
             <div className={container + " d-flex align-content-center"}>
                 <div className="dialog-content">
                     <h3 className="text-center">Confirm Transaction</h3>
                     <table className="data w-100 mt-5">
                         <tbody>
-                            <tr>
-                                <td>Operation Type</td>
-                                <td className="text-right text-secondary">{this.props.data.showTransfer ? "Transfer" : "Withdraw"}</td>
-                            </tr>
-                            <tr>
-                                <td>Destination Account</td>
-                                <td className="text-right text-secondary">{this.props.data.showTransfer ? this.props.data.destination : this.props.data.issuer}</td>
-                            </tr>
-                            <tr>
-                                <td>Amount</td>
-                                <td className="text-right text-secondary">{this.props.data.amount} <SymbolToken name={this.props.data.asset} showIcon={false}/></td>
-                            </tr>
-                            <tr>
-                                <td colSpan={2}>{this.props.data.showTransfer ? "Memo" : "Beneficiary Address"}</td>
-                            </tr>
-                            <tr>
-                                <td colSpan={2} className="text-secondary pl-3">{this.props.data.memo}</td>
-                            </tr>
-                            <tr>
-                                <td>Platform Fee</td>
-                                <td className="text-right text-secondary">{this.props.data.fee.amount} {this.props.data.fee.asset}</td>
-                            </tr>
+                            { data.type ?
+                                <tr>
+                                    <td>Operation Type</td>
+                                    <td className="text-right text-secondary">{data.type}</td>
+                                </tr>
+                                : null
+                            }
+                            { data.destination ?
+                                <tr>
+                                    <td>Destination Account</td>
+                                    <td className="text-right text-secondary">{data.destination}</td>
+                                </tr>
+                                : null
+                            }
+                            { data.amount ?
+                                <tr>
+                                    <td>Amount</td>
+                                    <td className="text-right text-secondary">{data.amount.amount} <SymbolToken name={data.amount.asset} showIcon={false} withLink={false} /></td>
+                                </tr>
+                                : null
+                            }
+                            { data.memo ?
+                                <React.Fragment>
+                                    <tr>
+                                        <td colSpan={2}>{data.memo.type}</td>
+                                    </tr>
+                                    <tr>
+                                        <td colSpan={2} className="text-secondary pl-3">{data.memo.memo}</td>
+                                    </tr>
+                                </React.Fragment>
+                                : null
+                            }
+                            { data.fee ?
+                                <tr>
+                                    <td>Platform Fee</td>
+                                    <td className="text-right text-secondary">{data.fee.amount} {data.fee.asset}</td>
+                                </tr>
+                                : null
+                            }
                         </tbody>
                     </table>
+                    {
+                        !has_fund ?
+                            <span className="text-danger small float-right">* Insufficient Fund</span>
+                            : null
+                    }
                     <div className="d-flex justify-content-around mt-5">
                         <button className="cancel-btn" onClick={this.props.cancel} disabled={this.state.processing}>CANCEL</button>
-                        <button disabled={this.state.processing} onClick={() => {
+                        <button disabled={this.state.processing || !has_fund} onClick={() => {
                             this.setState({processing: true})
                             this.props.submit()
                         }}>{this.state.processing ? <Loader /> : "Accept"}</button>
