@@ -4,6 +4,10 @@ import Exchange from './components/exchange.jsx';
 import Fund from './components/fund.jsx';
 import Message from './components/message.jsx';
 import Leaderboard from './components/leaderboard_full.jsx';
+import Landing from './pages/main/landing.jsx'
+import Technology from './pages/main/technology.jsx'
+import ExportKey from './components/export_key.jsx'
+import DiceGame from './pages/games/dice.jsx'
 
 import { createStore, applyMiddleware, compose } from 'redux'
 import { Provider } from 'react-redux'
@@ -15,11 +19,27 @@ import { Router, Route, Switch } from 'react-router-dom'
 
 import { injectGlobal } from 'emotion'
 import globalcss from './components/global-css.js'
-import createHashHistory from 'history/createHashHistory'
-import createBrowserHistory from 'history/createBrowserHistory'
+import { createHashHistory, createBrowserHistory } from 'history'
 import ReactGA from 'react-ga';
  
-ReactGA.initialize('UA-114919036-3');
+ReactGA.initialize(window.isApp ? 'UA-114919036-4': 'UA-114919036-3');
+ReactGA.set({ checkProtocolTask: null })
+
+if (typeof document.hidden !== "undefined") { // Opera 12.10 and Firefox 18 and later support 
+  window.hidden = "hidden";
+} else if (typeof document.msHidden !== "undefined") {
+  window.hidden = "msHidden";
+} else if (typeof document.webkitHidden !== "undefined") {
+  window.hidden = "webkitHidden";
+}
+
+window.addEventListener('keyboardDidShow', function () {
+	document.getElementById('app').classList.add("keyboard-show")
+});
+
+window.addEventListener('keyboardDidHide', function () {
+	document.getElementById('app').classList.remove("keyboard-show")
+});
 
 // , applyMiddleware(logger)
 
@@ -192,6 +212,65 @@ injectGlobal`
 	.cursor-pointer {
 		cursor: pointer;
 	}
+
+	.keyboard-show {
+		.exchange-bottom {
+			display: none !important;
+		}
+
+		.mobile-content {
+			margin-bottom: 0;
+		}
+	}
+	
+	.blue-btn {
+		-moz-box-shadow:inset 0px 1px 0px 0px #54a3f7;
+		-webkit-box-shadow:inset 0px 1px 0px 0px #54a3f7;
+		box-shadow:inset 0px 1px 0px 0px #54a3f7;
+		background:-webkit-gradient(linear, left top, left bottom, color-stop(0.05, #007dc1), color-stop(1, #0061a7));
+		background:-moz-linear-gradient(top, #007dc1 5%, #0061a7 100%);
+		background:-webkit-linear-gradient(top, #007dc1 5%, #0061a7 100%);
+		background:-o-linear-gradient(top, #007dc1 5%, #0061a7 100%);
+		background:-ms-linear-gradient(top, #007dc1 5%, #0061a7 100%);
+		background:linear-gradient(to bottom, #007dc1 5%, #0061a7 100%);
+		filter:progid:DXImageTransform.Microsoft.gradient(startColorstr='#007dc1', endColorstr='#0061a7',GradientType=0);
+		background-color:#007dc1;
+		-moz-border-radius:3px;
+		-webkit-border-radius:3px;
+		border-radius:3px;
+		border:1px solid #124d77;
+		display:inline-block;
+		cursor:pointer;
+		color:#ffffff;
+		font-family:Arial;
+		font-size:13px;
+		padding:6px 24px;
+		text-decoration:none;
+		text-shadow:0px 1px 0px #154682;
+	}
+	.blue-btn:hover {
+		background:-webkit-gradient(linear, left top, left bottom, color-stop(0.05, #0061a7), color-stop(1, #007dc1));
+		background:-moz-linear-gradient(top, #0061a7 5%, #007dc1 100%);
+		background:-webkit-linear-gradient(top, #0061a7 5%, #007dc1 100%);
+		background:-o-linear-gradient(top, #0061a7 5%, #007dc1 100%);
+		background:-ms-linear-gradient(top, #0061a7 5%, #007dc1 100%);
+		background:linear-gradient(to bottom, #0061a7 5%, #007dc1 100%);
+		filter:progid:DXImageTransform.Microsoft.gradient(startColorstr='#0061a7', endColorstr='#007dc1',GradientType=0);
+		background-color:#0061a7;
+	}
+	.blue-btn:active {
+		position:relative;
+		top:1px;
+	}	
+
+	.grecaptcha-badge {
+		display: none;
+	}
+
+	.__react_component_tooltip {
+		white-space: normal !important;
+		line-height: 20px;
+	}
 `
 
 const store = createStore(reducer, compose(applyMiddleware(thunk)))
@@ -202,26 +281,43 @@ if (window.isApp) {
 } else {
 	history = createBrowserHistory()
 }
+
 history.listen((location, action) => {
 	//console.log("history change ", location.pathname);
 	ReactGA.set({ page: location.pathname });
 	ReactGA.pageview(location.pathname);
 });
 
+if (/Android|iPad|iPhone|iPod/i.test(navigator.userAgent)) {
+	$(document).on('focus', 'input, textarea', function() {
+		document.getElementById('app').classList.add('kb-opened')
+	});
+
+	$(document).on('blur', 'input, textarea', function() {
+		document.getElementById('app').classList.remove('kb-opened')
+	});
+}
 
 class Container extends React.Component {
   render () {
     return (
     <Provider store={store}>
-				<Router history={history}>
-        <Switch>
-					<Route exact path="/" component={Exchange} />
-					<Route exact path="/exchange" component={Exchange} />
-					<Route path="/exchange/wallets" component={Fund} />
-					<Route exact path="/exchange/message" component={Message} />
-					<Route exact path="/leaderboard" component={Leaderboard} />
-        </Switch>
-      </Router>
+		<Router history={history}>
+			<Switch>
+				<Route exact path="/" component={window.isApp ? Exchange : Landing} />
+				<Route exact path="/technology" component={Technology} />
+				<Route exact path="/:net" component={Exchange} />
+				<Route exact path="/:net/dice" component={DiceGame} />
+				<Route exact path="/:net/exchange" component={Exchange} />
+				<Route exact path="/:net/exchange/:ticker" component={Exchange} />
+				<Route exact path="/:net/wallets" component={Fund} />
+				<Route exact path="/:net/message" component={Message} />
+				<Route exact path="/:net/export_key" component={ExportKey} />
+				<Route exact path="/:net/leaderboard" component={Leaderboard} />
+
+				<Route component={window.isApp ? Exchange : Landing}/>
+			</Switch>
+      	</Router>
     </Provider>);
   }
 }
