@@ -1,19 +1,24 @@
 import lodash from 'lodash';
-import API from "../../api.jsx"
-import SortedSet from 'js-sorted-set'
-import QuantaClient from "@quantadex/quanta_js"
+import React from 'react';
 import ApplicationApi from "../../common/api/ApplicationApi";
+import WalletApi from "../../common/api/WalletApi";
 import { Apis } from "@quantadex/bitsharesjs-ws";
-import { Price, Asset, FillOrder, LimitOrderCreate, LimitOrder } from "../../common/MarketClasses";
-import { PrivateKey, PublicKey, Aes, key, ChainStore } from "@quantadex/bitsharesjs";
+import { FillOrder, LimitOrder } from "../../common/MarketClasses";
+import { PrivateKey, ChainStore } from "@quantadex/bitsharesjs";
 import { createLimitOrderWithPrice, createLimitOrder2, cancelOrder, signAndBroadcast } from "../../common/Transactions";
-import { aggregateOrderBook, convertHistoryToOrderedSet } from "../../common/PriceData";
-import ReactGA from 'react-ga';
+import { validMarketPair, getBaseCounter, aggregateOrderBook, convertHistoryToOrderedSet } from "../../common/PriceData";
 import { toast } from 'react-toastify';
+import ReactGA from 'react-ga';
 import CONFIG from '../../config.js'
+import Utils from "../../common/utils.js";
+import {setItem} from "../../common/storage.js"
+import * as QuantaApi from './quanta-api'
+import * as OrderHistory from './order_history.jsx';
 
 export const INIT_DATA = 'INIT_DATA';
+export const USER_DATA = 'USER_DATA';
 export const LOGIN = 'LOGIN';
+export const LOGOUT = 'LOGOUT';
 export const UPDATE_ACCOUNT = 'UPDATE_ACCOUNT';
 export const UPDATE_FEE = 'UPDATE_FEE';
 export const SET_MARKET_QUOTE = 'SET_MARKET_QUOTE';
@@ -25,105 +30,168 @@ export const UPDATE_USER_ORDER = 'UPDATE_USER_ORDER';
 export const UPDATE_TRADES = 'UPDATE_TRADES'
 export const UPDATE_DIGITS = 'UPDATE_DIGITS'
 export const UPDATE_BLOCK_INFO = 'UPDATE_BLOCK_INFO';
-export const TOGGLE_LEFT_PANEL = 'TOGGLE_LEFT_PANEL';
-export const TOGGLE_RIGHT_PANEL = 'TOGGLE_RIGHT_PANEL';
-export const UPDATE_NETWORK = 'UPDATE_NETWORK';
+export const UPDATE_STORAGE = 'UPDATE_STORAGE';
 export const LOAD_FILLED_ORDERS = 'LOAD_FILLED_ORDERS';
-
-export const TOGGLE_FAVORITE_LIST = 'TOGGLE_FAVORITE_LIST';
+export const TOGGLE_CONNECT_DIALOG = 'TOGGLE_CONNECT_DIALOG';
+export const TOGGLE_BUY_QDEX_DIALOG = 'TOGGLE_BUY_QDEX_DIALOG';
 export const INIT_BALANCE = 'INIT_BALANCE'
 export const UPDATE_OPEN_ORDERS = 'UPDATE_OPEN_ORDERS'
+export const WEBSOCKET_STATUS = 'WEBSOCKET_STATUS'
 
-export const toggleFavoriteList = pair => ({
-	type: TOGGLE_FAVORITE_LIST,
-	pair
-})
-var markets = null;
 
-export function getMarketQuotes() {
-	return function(dispatch) {
-
-	} 
-}
-
-export function initBalance() {
-	return function(dispatch) {
-	}
-}
-
-function getBaseCounter(market) {
-	const parts = market.split("/")
-	return {
-		base: assetsBySymbol[parts[0]],
-		counter: assetsBySymbol[parts[1]]
-	}
-}
-
-export function GetName(id) {
+export function GetAccount(id) {
 	return Apis.instance().db_api().exec("get_accounts", [[id]]).then(e => {
-		return e[0].name
+		return e[0]
 	})
 }
 
-export function buyTransaction(market, price, amount) {
+export function GetAccountBalances(id, assets=[]) {
+	return Apis.instance().db_api().exec("get_account_balances", [id, assets]).then(e => {
+		return e
+	})
+}
+
+export function GetBlock(id) {
+	return Apis.instance().db_api().exec("get_block", [id]).then(e => {
+		return e
+	})
+}
+
+export function GetChainId() {
+	return Apis.instance().db_api().exec("get_chain_id", []).then(e => {
+		return e
+	})
+}
+
+function notify_success(toastId, msg) {
+	toast.update(toastId, {
+		render: msg,
+		type: toast.TYPE.SUCCESS,
+		autoClose: 2000,
+		position: toast.POSITION.TOP_CENTER,
+		pauseOnFocusLoss: false,
+		pauseOnHover: false
+	});
+}
+
+function notify_failed(toastId, msg) {
+	toast.update(toastId, {
+		render: msg,
+		type: toast.TYPE.ERROR,
+		autoClose: 2000,
+		position: toast.POSITION.TOP_CENTER,
+		pauseOnFocusLoss: false,
+		pauseOnHover: false
+	});
+}
+
+function toastMsg(label, success, e) {
+	const msg = (<div>
+		<span>{label}</span><br />
+		<span>{success ? "Order ID: " + e.trx.operation_results[0][1] :
+			"Failed order: " + (e.message.toLowerCase().includes("insufficient balance") ? "Insufficient Balance" : "Unable to place order: " + e.message)}</span>
+	</div>)
+	return msg
+}
+
+export function buyTransaction(market, price, amount, fill_or_kill = false, mobile_nav = undefined) {
 	return (dispatch, getState) => {
 		var {base, counter} = getBaseCounter(market)
+
+		const asset = market.split('/')[0].split('0X')
+		const label = asset[0] + (asset[1] ? '0x' + asset[1].substr(0,4) : "") + " " + amount + " @ " + price
+		const toastId = toast("BUYING " + label, { autoClose: false, position: toast.POSITION.TOP_CENTER });
+
 		var user_id = getState().app.userId;
-
 		const pKey = PrivateKey.fromWif(getState().app.private_key);
-		// console.log(pKey, assets[base.id], "price=",price, "amount=", amount, user_id);
-
-		const order = createLimitOrderWithPrice(user_id, true, window.assets, base.id, counter.id, price, amount)
-
-		// console.log("order prepare", order);
+		const order = createLimitOrderWithPrice(user_id, true, window.assets, base.id, counter.id, price, amount, fill_or_kill)
 		const tr = createLimitOrder2(order)
+
 		return signAndBroadcast(tr, pKey)
 			.then((e) => {
-				// console.log("order result ", e);
-				return e[0]
+				const msg = toastMsg("BUY " + label, true, e[0])
+				dispatch(updateUserData())
+				if (asset[0] == "QDEX") {
+					if (mobile_nav) {
+						mobile_nav("trade")
+					} else {
+						dispatch({
+							type: TOGGLE_BUY_QDEX_DIALOG,
+							data: false
+						})
+					}
+				}
+				setTimeout(() => {
+					notify_success(toastId, msg)
+				}, 0)
 			}).catch((e) => {
-				throw e
+				if (e.message.includes("less than required")) {
+					if (mobile_nav) {
+						mobile_nav("buy_qdex")
+					} else {
+						dispatch({
+							type: TOGGLE_BUY_QDEX_DIALOG,
+							data: true
+						})
+					}
+				}
+				const msg = toastMsg("BUY " + label, false, e)
+				notify_failed(toastId, msg)  
+				Rollbar.error("Buy Failed", e);
+				return e
 			})
 	}
 }
 
-export function sellTransaction(market, price, amount) {
+export function sellTransaction(market, price, amount, mobile_nav = undefined) {
 	return (dispatch, getState) => {
 		var { base, counter } = getBaseCounter(market)
+
+		const asset = market.split('/')[0].split('0X')
+		const label = asset[0] + (asset[1] ? '0x' + asset[1].substr(0,4) : "") + " " + amount + " @ " + price
+		const toastId = toast("SELLING " + label, { autoClose: false, position: toast.POSITION.TOP_CENTER });
+	
 		var user_id = getState().app.userId;
-
 		const pKey = PrivateKey.fromWif(getState().app.private_key);
-		// console.log(pKey, assets[base.id], "price=", price, "amount=", amount, user_id);
-
 		const order = createLimitOrderWithPrice(user_id, false, window.assets, base.id, counter.id, price, amount)
-		
-		// console.log("order prepare", order);
 		const tr = createLimitOrder2(order)
+
 		return signAndBroadcast(tr, pKey)
 			.then((e) => {
-				// console.log("order result ", e);
-				return e[0]
+				const msg = toastMsg("SELL " + label, true, e[0])
+				notify_success(toastId, msg)
+				dispatch(updateUserData())
 			}).catch((e) => {
-				throw e
+				if (e.message.includes("less than required")) {
+					if (mobile_nav) {
+						mobile_nav("buy_qdex")
+					} else {
+						dispatch({
+							type: TOGGLE_BUY_QDEX_DIALOG,
+							data: true
+						})
+					}
+				}
+				const msg = toastMsg("SELL " + label, false, e)
+				notify_failed(toastId, msg)
+				Rollbar.error("Sell Failed", e);
 			})
 	}
 
 }
 
-export const cancelTransaction = (market, order_id) => {
+export const cancelTransaction = (order_id) => {
 	return (dispatch, getState) => {
-		var { base, counter } = getBaseCounter(market)
 		var user_id = getState().app.userId;
-	
 		const pKey = PrivateKey.fromWif(getState().app.private_key);
-		// console.log(pKey, assets[base.id], user_id);
-	
 		const order = cancelOrder(user_id, order_id)
-	
-		// console.log("cancel order", order);
+		
 		return signAndBroadcast(order, pKey)
 			.then((e) => {
-				// console.log("order result ", e);
+				dispatch(updateUserData())
+			}).catch((e) => {
+				Rollbar.error("Cancel Failed", e);
+				throw e
 			})
 	}
 }
@@ -139,11 +207,11 @@ export const transferFund = (data) => {
 			broadcast: true,
 			encrypt_memo: false,
 		}).then((tr) => {
-			// console.log(tr);
-			return signAndBroadcast(tr, PrivateKey.fromWif(getState().app.private_key))
-		}).then(e => {
-			switchTicker(getState().app.currentTicker)
+			return signAndBroadcast(tr, PrivateKey.fromWif(getState().app.private_key)).then(() => {
+					dispatch(updateUserData())
+				})
 		}).catch(e => {
+			Rollbar.error("Transfer Failed", e);
 			throw e
 		})
 	}
@@ -157,144 +225,72 @@ export const withdrawVesting = (data) => {
 			amount: data.amount,
 			asset: data.asset
 		}).then((tr) => {
-			return signAndBroadcast(tr, PrivateKey.fromWif(getState().app.private_key))
-		}).then(e => {
-			switchTicker(getState().app.currentTicker)
+			return signAndBroadcast(tr, PrivateKey.fromWif(getState().app.private_key)).then(() => {
+				dispatch(updateUserData())
+			})
 		}).catch(e => {
+			Rollbar.error("Vesting Failed", e);
 			throw e
 		})
 	}
 }
 
-var initAPI = false;
-var initUser = false;
-
-export var dataSize = 100;
-
-function getAvgtime() {
-	const avgTime = blockTimes.reduce((previous, current, idx, array) => {
-		return previous + current[1] / array.length;
-	}, 0);
-	return avgTime;
-}
-const blockTimes = [];
-var previousTime = null;
-
-function updateChainState(dispatch) {
-	const object = ChainStore.getObject("2.1.0");
-
-	if (object==null) {
-		return;
-	}
-
-	const timestamp = timeStringToDate(object.get("time"))
-	
-	if (previousTime != null) {
-		blockTimes.push([
-			object.get("head_block_number"),
-			(timestamp - previousTime) / 1000
-		]);
-	}
-
-	previousTime = timestamp;
-
-	//console.log('update', timestamp, object.get("head_block_number"), Math.round(getAvgtime()*1000))
-
-	dispatch({
-		type: UPDATE_BLOCK_INFO,
-		data: {
-			blockNumber: object.get("head_block_number"),
-			blockTime: Math.round(getAvgtime() * 1000)
-		}
-	})
-}
-
-
-function timeStringToDate(block_time) {
-	if (!/Z$/.test(block_time)) {
-		block_time += "Z";
-	}
-	return new Date(block_time);
-}
-
-
-var dynamicGlobal = null;
-
-function onUpdate(dispatch) {
-
-}
-
-function validMarketPair(a, b) {
-	const pair = [a, b]
-	const base = assets[a].symbol
-	const counter = assets[b].symbol
-	const symbolA = [base, counter]
-
-	const foundA = window.marketsHash[symbolA.join("/")]
-
-	return foundA ? pair : pair.reverse()
-}
-
-async function processOrderHistory(data, userId) {
-	const my_history = []
-	const cancels = []
-	var limitOrders = []
-	
-	data.filter((op) => op.operation_type == 2).map(item => {
-		cancels.push(item.operation_history.op_object.order)
-	})
-
-	if (cancels.length > 0) {
-		limitOrders = await fetch(CONFIG.SETTINGS[window.currentNetwork].API_PATH + `/account?operation_type=1&account_id=${userId}&size=100&filter_field=operation_history__operation_result&filter_value=${cancels.join(',')}`)
-			.then(e => e.json())
-			.then(e => {
-				return lodash.keyBy(e, (o) => o.operation_history.operation_result.split(',')[1].replace(/"/g, '').replace(']', ''))
+export const withdrawGenesis = (data) => {
+	return (dispatch, getState) => {
+		return ApplicationApi.balance_claim({ 
+			account: getState().app.userId,
+			public_key: getState().app.publicKey,
+			balance: data.balance_id,
+			amount: data.amount,
+			asset: data.asset
+		}).then((tr) => {
+			return signAndBroadcast(tr, PrivateKey.fromWif(getState().app.private_key)).then(() => {
+				dispatch(updateUserData())
 			})
+		}).catch(e => {
+			Rollbar.error("Genesis Failed", e);
+			throw e
+		})
 	}
-	data.forEach((filled) => {
-		if (filled.operation_type == 2 && filled.operation_history.operation_result == "[0,{}]") {
-			return
-		}
-		let baseId
-		var order
-		const type = filled.operation_type
-		let op = filled.operation_history
-		op.time = filled.block_data.block_time
-		op.block_num = filled.block_data.block_num
+}
 
-		if (type == 2) {
-			let limitOp = limitOrders[op.op_object.order].operation_history.op_object
-			limitOp.sell_price = {base: limitOp.amount_to_sell, quote: limitOp.min_to_receive}
-			limitOp.id = op.op_object.order
-			limitOp.time = op.time
-			
-			baseId = validMarketPair(limitOp.amount_to_sell.asset_id, limitOp.min_to_receive.asset_id)[0]
-			order = new LimitOrder(
-				limitOp,
-				window.assets,
-				baseId
-			);
-		} else {
-			baseId = validMarketPair(op.op_object.fill_price.base.asset_id, op.op_object.fill_price.quote.asset_id)[0]
-			op.op = [{}, op.op_object]
-			order = new FillOrder(
-				op,
-				window.assets,
-				baseId
-			);
-		}
-		my_history.push(order)
-	})
-	return my_history
+export const accountUpgrade = () => {
+	return (dispatch, getState) => {
+		return ApplicationApi.account_upgrade({ 
+			account: getState().app.userId
+		}).then((tr) => {
+			return signAndBroadcast(tr, PrivateKey.fromWif(getState().app.private_key)).then(() => {
+				let data = {
+					id: getState().app.userId,
+					name: getState().app.name,
+					publicKey: getState().app.publicKey,
+					lifetime: true
+				}
+				dispatch({
+					type: UPDATE_ACCOUNT,
+					data: data
+				})
+				dispatch(updateUserData())
+			})
+		}).catch(error => {
+			console.log(error)
+			if (error.message.includes("Insufficient Balance")) {
+				throw "Insufficient Balance"
+			} else {
+				Rollbar.error("Upgrade Failed", error);
+				throw "Server error. Please try again"
+			}
+		})
+	}
 }
 
 export const loadOrderHistory = (page) => {
 	return (dispatch, getState) => {
 		const userId = getState().app.userId
-		return fetch(CONFIG.SETTINGS[window.currentNetwork].API_PATH + `/account?filter_field=operation_type&filter_value=2,4&size=${dataSize}&from_=${page * dataSize}&account_id=${userId}`)
+		return fetch(CONFIG.getEnv().API_PATH + `/account?filter_field=operation_type&filter_value=2,4&size=${dataSize}&from_=${page * dataSize}&account_id=${userId}`, { mode: "cors" })
 			.then(e => e.json())
 			.then((filled) => {
-				return processOrderHistory(filled, userId)
+				return OrderHistory.processOrderHistory(filled, userId)
 			}).then(my_history => {
 				dispatch({
 					type: LOAD_FILLED_ORDERS,
@@ -304,235 +300,339 @@ export const loadOrderHistory = (page) => {
 	}
 }
 
-var disconnect_notified
+export const rollDice = (amount, asset, bet) => {
+	return (dispatch, getState) => {
+		return ApplicationApi.roll_dice({ 
+			account: getState().app.userId,
+			amount,
+			asset,
+			bet,
+			numbers: []
+		}).then((tr) => {
+			return signAndBroadcast(tr, PrivateKey.fromWif(getState().app.private_key))
+				.then((res) => {
+					// console.log("Call order update success!");
+					return tr.id()
+				}).catch((e) => {
+					console.log(e)
+					if (e.message.includes("less than required")) {
+						dispatch({
+							type: TOGGLE_BUY_QDEX_DIALOG,
+							data: true
+						})
+						throw "Insufficient QDEX"
+					}
+					if (e.message.includes("Insufficient Balance")) throw "Insufficient Fund"
+					throw e.message
+				})
+		}).catch(error => {
+			console.log(error)
+			throw error
+		})
+	}
+}
+
+export const AccountLogin = (private_key) => {
+	return (dispatch, getState) => {
+		const pKey = PrivateKey.fromWif(private_key);
+		let publicKey = pKey.toPublicKey().toString()
+
+		return Apis.instance()
+		.db_api()
+		.exec("get_key_references", [[publicKey]])
+		.then(vec_account_id => {
+			// console.log("get_key_references ", vec_account_id);
+			if (vec_account_id[0].length == 0) {
+				throw "No account for public key: " + publicKey
+			}
+
+			const dedup_ids = []
+			for (let id of vec_account_id[0]) {
+				if (!dedup_ids.includes(id)) dedup_ids.push(id)
+			}
+
+			if (dedup_ids.length == 1) {
+				dispatch(ConnectAccount(dedup_ids[0], private_key))
+			} else {
+				return Apis.instance().db_api().exec("get_accounts", [dedup_ids]).then(e => {
+					return e
+				})
+			}
+			
+			
+		}).catch(error => {
+			console.log(error)
+			if (typeof error == "string") {
+				throw error
+			} else {
+				Rollbar.error("Login Failed", error);
+				throw "Server error. Please try again."
+			}
+		})
+	}
+}
+
+export const ConnectAccount = (account_id, private_key) => {
+	return (dispatch, getState) => {
+		return Apis.instance()
+		.db_api()
+		.exec("get_objects", [[account_id]])
+		.then((data) => {
+			// console.log("get account ", data);
+			let lifetime_member = data[0].membership_expiration_date === "1969-12-31T23:59:59"
+			dispatch({
+				type: UPDATE_ACCOUNT,
+				data: {...data[0], lifetime: lifetime_member}
+			})
+			dispatch({
+				type: LOGIN,
+				private_key: private_key
+			})
+
+			setItem("name", data[0].name)
+			setItem("id", data[0].id)
+			setItem("publicKey", data[0].active.key_auths[0][0])
+			setItem("lifetime", lifetime_member)
+			setItem("env", window.location.pathname.startsWith("/testnet") ? "testnet" : "mainnet")
+			if (window.isApp) {
+				setItem("private_key", private_key)
+			}
+
+			Rollbar.configure({
+				payload: {
+				  person: {
+					id: data[0].id,
+					username: data[0].name
+				  }
+				}
+			  });
+
+		}).then(e => {
+			dispatch(updateUserData())
+			dispatch({
+				type: TOGGLE_CONNECT_DIALOG,
+				data: false
+			})
+			return true
+		}).catch(error => {
+			console.log(error)
+			if (error.message.includes("Bad Cast")) {
+				throw "No account for public key: " + publicKey
+			} else {
+				Rollbar.error("Connect Failed", error);
+				throw "Server error. Please try again."
+			}
+		})
+	}
+}
+
+export function getQdexAsks(counter) {
+	return function () {
+		const qdex_id = window.assetsBySymbol.QDEX.id
+		return Apis.instance().db_api().exec("get_order_book", [counter.id, qdex_id, 50]).then((ob) => {
+			return aggregateOrderBook([], ob.asks, window.assets[qdex_id].precision).asks
+		})
+	}
+}
+
+window.initAPI = false;
+export var dataSize = 100;
+
 export function switchTicker(ticker, force_init=false) {
-	Apis.setAutoReconnect(true)
 	// send GA
 	ReactGA.set({ page: "exchange/" + ticker });
 	ReactGA.pageview("exchange/" + ticker);
 	// console.log("Switch ticker ", ticker);
 
 	return function (dispatch,getState) {
-		var publicKey = null
-		if (getState().app.private_key !== null) {
-			const pKey = PrivateKey.fromWif(getState().app.private_key);
-			publicKey = pKey.toPublicKey().toString()
+		async function update() {
+			dispatch({ type: UPDATE_TICKER, data: ticker })
+			if (window.initAPI == false || force_init) {
+				Apis.setRpcConnectionStatusCallback(
+					(status) => {
+						// console.log("ws status changed: ", status);
+						if (status === "reconnect")
+							ChainStore.resetCache(false);
+						if (status === "closed") {
+							dispatch({ type: "WEBSOCKET_STATUS", data: "Closed" })
+							setTimeout(() => {
+								if (!document[window.hidden]) QuantaApi.ConnectAsync(dispatch)
+							}, 1000)
+						}
+						if (status === "open") {
+							dispatch({ type: "WEBSOCKET_STATUS", data: null })
+						}
+					}
+				);
+
+				const res = await QuantaApi.ConnectAsync(dispatch)
+
+				await QuantaApi.UpdateGlobalPropertiesAsync()
+				await fetchAndSubscribeTickerAsync(ticker, dispatch)
+
+				// Updates user in case they are already logged in.
+				Rollbar.configure({
+					payload: {
+						person: {
+						id: getState().app.userId,
+						username: getState().app.name
+						}
+					}
+				});
+				window.initAPI = true;
+				// console.log("initialized update");
+			} else {
+				await fetchAndSubscribeTickerAsync(ticker, dispatch)
+			}
+		}
+		update().catch((e)=>{
+			Rollbar.error("update ticker exception: ", e);
+		})
+	}
+}
+
+async function fetchAndSubscribeTickerAsync(ticker, dispatch) {
+	var {base, counter} = getBaseCounter(ticker)
+	if (!base || !counter) 
+		return
+
+	// update fees
+	const tmpOrder = createLimitOrderWithPrice("1.2.0", true, window.assets, base.id, counter.id, 1, 1)
+	const tr = createLimitOrder2(tmpOrder)
+	await tr.set_required_fees()
+	dispatch({
+		type: UPDATE_FEE,
+		data: tr.operations[0][1].fee
+	})
+	
+	var fetchTime = new Date()
+	Apis.instance().db_api().exec("subscribe_to_market", [(data) => {
+		// console.log("Got a market change ", ticker, base, counter, data);
+		const curr_ticker = getBaseCounter(ticker)
+
+		async function subscribeToTicker() {
+			if (base.id === curr_ticker.base.id && counter.id === curr_ticker.counter.id) {
+				if (new Date() - fetchTime > 50) {
+					await updateFetchDataAsync(ticker)
+					fetchTime = new Date()
+
+					if (Apis.instance().streamCb) {
+						await Apis.instance().streamCb()
+					}			
+				}
+			}
 		}
 
+		subscribeToTicker().then(() => {}).catch((e) => {
+			Rollbar.error("subscribe exception caught:", e);
+		})
+	}, base.id, counter.id])
+
+	async function updateFetchDataAsync(ticker, first=false) {
+		const data = await QuantaApi.fetchDataAsync(ticker)
+
+		window.USD_value = data.USD_value
 		dispatch({
-			type: UPDATE_TICKER,
-			data: ticker
+			type: SET_MARKET_QUOTE,
+			data: [data.marketData, data.USD_value]
 		})
 
-		if (initAPI == false || force_init) {
-			Apis.instance(CONFIG.SETTINGS[window.currentNetwork].WEBSOCKET_PATH, true, 3000, { enableOrders: true }).init_promise.then((res) => {
-				// console.log("connected to:", res[0].network, publicKey);
-
-				// Apis.instance().db_api().exec("set_subscribe_callback", [onUpdate, true]);
-				initAPI = true;
-
-				ChainStore.subscribers.clear()
-				ChainStore.init(false).then(() => {
-					ChainStore.subscribe(updateChainState.bind(this, dispatch));
-				});
-			})
-			.then((e) => {
-				Apis.instance().db_api().exec("list_assets", ["A", 100]).then((assets) => {
-					// console.log("assets ", assets);
-					window.assets = lodash.keyBy(assets, "id")
-					window.assetsBySymbol = lodash.keyBy(assets, "symbol")
-					return assets;
-				}).then((e) => {
-					action(ticker)
-				});
-			})
-			
-		} else {
-				action(ticker)
-		}
-
-		function action(ticker) {
-			var {base, counter} = getBaseCounter(ticker)
-
-			const tmpOrder = createLimitOrderWithPrice("1.2.0", true, window.assets, base.id, counter.id, 1, 1)
-			const tr = createLimitOrder2(tmpOrder)
-			tr.set_required_fees().then(e => {
-				dispatch({
-					type: UPDATE_FEE,
-					data: tr.operations[0][1].fee
-				})
-			})
-
-			async function fetchData(ticker, first=false) {
-				var {base, counter} = getBaseCounter(ticker)
-				try {
-					await fetch(CONFIG.SETTINGS[window.currentNetwork].MARKETS_JSON).then(e => e.json())
-					.then(async (e) => {
-						// save for later
-						markets = e;
-						window.markets = markets.markets
-						window.marketsHash = lodash.keyBy(markets.markets, "name")
-						
-						var marketData = [];
-						var USD_value = {}
-						// console.log("json ", markets.markets);
-
-						for (const market of markets.markets) {
-							var { base, counter } = getBaseCounter(market.name);
-							if(!base || !counter) continue
-							const data = await Promise.all([Apis.instance()
-								.db_api()
-								.exec("get_ticker", [counter.id, base.id]),
-							Apis.instance()
-								.db_api()
-								.exec("get_24_volume", [counter.id, base.id])])
-
-							marketData.push({
-								name: market.name,
-								last: data[0].latest,
-								base_volume: data[1].base_volume,
-								quote_volume: data[1].quote_volume
-							})
-							if (counter.symbol == 'USD') {
-								USD_value[base.id] = data[0].latest
-								USD_value[counter.id] = 1
-							}
-						}
-
-						dispatch({
-							type: SET_MARKET_QUOTE,
-							data: [marketData, USD_value]
-						})
-					}).then(e => {
-						if (!initUser && getState().app.private_key !== null) {
-							const pKey = PrivateKey.fromWif(getState().app.private_key);
-							publicKey = pKey.toPublicKey().toString()
-
-							Apis.instance()
-							.db_api()
-							.exec("get_key_references", [[publicKey]])
-							.then(vec_account_id => {
-								// console.log("get_key_references ", vec_account_id[0][0]);
-		
-								return Apis.instance()
-									.db_api()
-									.exec("get_objects", [[vec_account_id[0][0]]])
-									.then((data) => {
-										// console.log("get account ", data);
-										dispatch({
-											type: UPDATE_ACCOUNT,
-											data: data[0]
-										})
-									})
-		
-							}).then(e => {
-								initUser = true
-								fetchData(ticker)
-							})
-						}
-					})
-				} catch(e) {
-					console.log(e)
-					if (!disconnect_notified) {
-						disconnect_notified = toast.error("Lost connection to server. Please refresh the page.", {
-							position: toast.POSITION.TOP_CENTER,
-							autoClose: false
-						});
-					}
-					
-					return
-				}		
-				
-				const trades = Apis.instance().history_api().exec("get_fill_order_history", [base.id, counter.id, 100]).then((filled) => {
-					// console.log("history filled ", filled);
-					const trade_history = convertHistoryToOrderedSet(filled, base.id)
-					//console.log("converted ", trade_history);
-					return trade_history
-				})
-
-				var my_trades = []
-
-				if (publicKey && getState().app.userId) {
-					const userId = getState().app.userId
-					my_trades = fetch(CONFIG.SETTINGS[window.currentNetwork].API_PATH + `/account?filter_field=operation_type&filter_value=2,4&size=${dataSize}&account_id=${userId}`).then(e => e.json())
-					.then((filled) => {
-						const my_history = processOrderHistory(filled, userId)
-						// console.log(my_history)
-						return my_history
-					})
-				}
-	
-				// selling counter
-				const orderBook = Apis.instance().db_api().exec("get_order_book", [counter.id, base.id, 50]).then((ob) => {
-					// console.log("ob  ", ob);
-					return aggregateOrderBook(ob.bids, ob.asks, window.assets[base.id].precision)
-				})
-
-				var account_data = [[],[]]
-				
-				if (publicKey && getState().app.userId) {
-					account_data = Apis.instance()
-					.db_api()
-					.exec("get_full_accounts", [[getState().app.userId], true])
-					.then(results => {
-						var orders = [];
-						results[0][1].limit_orders.forEach((ordered) => {
-							// search both market pairs to see what makes sense
-							// as the trading pair
-							let baseId = validMarketPair(ordered.sell_price.base.asset_id, ordered.sell_price.quote.asset_id)[0]
-							var order = new LimitOrder(
-								ordered,
-								window.assets,
-								baseId
-							);
-							
-							// console.log("ordered", order);
-							orders.push(order)
-						})
-						return [results, orders]
-					})
-				}
-
-				// console.log("Get all the data! for ", ticker);
-				return Promise.all([orderBook, trades, account_data, my_trades])
-					.then((data) => {
-						dispatch({
-							type: INIT_DATA,
-							data: {
-								orderBook: data[0],
-								trades: data[1],
-								filledOrders: data[3],
-								openOrders: data[2][1],
-								ticker: ticker,
-								accountData: data[2][0]
-							}
-						})
-						if (first) {
-							const ask_section = document.getElementById("ask-section")
-							ask_section.scrollTop = ask_section.scrollHeight;
-						}
-						
-					})					
+		dispatch({
+			type: INIT_DATA,
+			data: {
+				ticker: data.ticker,
+				orderBook: data.orderBook,
+				trades: data.trades,
 			}
-			var fetchTime = new Date()
-			Apis.instance().db_api().exec("subscribe_to_market", [(data) => {
-				// console.log("Got a market change ", base, counter, data);
-				const curr_ticker = getBaseCounter(getState().app.currentTicker)
-				
-				if (base.id === curr_ticker.base.id && counter.id === curr_ticker.counter.id) {
-					if (new Date() - fetchTime > 50) {
-						fetchData(ticker)
-						fetchTime = new Date()
-					}
-					
-					if (Apis.instance().streamCb) {
-						Apis.instance().streamCb()
-					}
-				}
-			}, base.id, counter.id])
-
-			fetchData(ticker, true)
+		})	
+		
+		if (first) {
+			const ask_section = document.getElementById("ask-section")
+			if (ask_section) ask_section.scrollTop = ask_section.scrollHeight;
 		}
 	}
 
+	await updateFetchDataAsync(ticker, true)
+}
+
+export function refreshData() {
+	return async function (dispatch, getState) {
+		if (getState().app.websocket_status) {
+			// console.log("Attempt to reconnect", getState().websocket_status);
+			const res = await QuantaApi.ConnectAsync(dispatch)
+		}
+		dispatch(switchTicker(getState().app.currentTicker))
+		dispatch(updateUserData())
+	}
+}
+
+export function updateUserData() {
+	return async function (dispatch, getState) {
+		const publicKey = getState().app.publicKey
+		const userId = getState().app.userId
+		// console.log("updateUserData", userId);	
+
+		if (!(publicKey && userId)) return
+		if (!Apis.instance().db_api() || getState().app.markets.length == 0) {
+			setTimeout(() => {
+				dispatch(updateUserData())
+			}, 500)
+			return
+		}
+		
+		const my_trades_promise = fetch(CONFIG.getEnv().API_PATH + `/account?filter_field=operation_type&filter_value=2,4&size=${dataSize}&account_id=${userId}`, { mode: "cors" }).then(e => e.json())
+		.then((filled) => {
+			const my_history = OrderHistory.processOrderHistory(filled, userId)
+			// console.log(my_history)
+			return my_history
+		})
+		
+		const account_data_promise = Apis.instance()
+		.db_api()
+		.exec("get_full_accounts", [[userId], true])
+		.then(results => {
+			var orders = [];
+			results[0][1].limit_orders.forEach((ordered) => {
+				// search both market pairs to see what makes sense
+				// as the trading pair
+				let baseId = validMarketPair(ordered.sell_price.base.asset_id, ordered.sell_price.quote.asset_id)[0]
+				var order = new LimitOrder(
+					ordered,
+					window.assets,
+					baseId
+				);
+				
+				// console.log("ordered", order);
+				orders.push(order)
+			})
+			return [results, orders]
+		}).then(e => {
+			let statistic_id = e[0][0][1].statistics.id
+			return Apis.instance().db_api().exec("get_objects", [[statistic_id]]).then(statistics=>{
+				return [...e, statistics[0].extensions.referral_fee_paid]
+			})
+		})
+
+		const shortAddress = WalletApi.getShortAddress(publicKey)
+		const genesis_balance_promise = Apis.instance().db_api().exec("get_balance_objects", [[shortAddress]]).then(e=>{
+			// console.log("balance obj????", e);
+			return e
+		})
+
+		return Promise.all([account_data_promise, my_trades_promise, genesis_balance_promise])
+			.then((data) => {
+				dispatch({
+					type: USER_DATA,
+					data: {
+						filledOrders: data[1],
+						openOrders: data[0][1],
+						accountData: data[0][0],
+						referral_fee: data[0][2],
+						genesis_balance: data[2],
+					}
+				})
+			}).catch(e=> {
+				Rollbar.error("Problems with updateUserData ", e)
+			})
+	}
 }
